refactor: extract Noise protocol into standalone fips-noise crate
Extract all Noise IK/XK/XX state machines, crypto helpers, and constants
from microfips-core/src/noise.rs (2616 lines) into a new workspace member:
crates/fips-noise/ (2629 lines).

This is a PoC for modularizing the FIPS Noise protocol layer:
- fips-noise is pure no_std, zero dependencies on microfips-core
- Exports 6 Noise types, all constants, helper functions (ecdh_pubkey,
  x_only_ecdh, aead_encrypt/decrypt, parity_normalize, etc.)
- microfips-core re-exports via 'pub use fips_noise as noise' for
  backward compatibility — all existing crate::noise:: imports work
- 37 unit tests in fips-noise, 125 in microfips-core (via re-export),
  134 in microfips-protocol — all pass

The crate can eventually be consumed directly by the upstream FIPS daemon
to replace its std-only src/noise/ module, achieving shared Noise code
across both implementations.
