+++ crates/microfips-core/Cargo.toml
@@ -9,6 +9,7 @@ autotests = false
 microfips-build = { path = "../microfips-build" }
 
 [dependencies]
+fips-noise = { path = "../fips-noise" }
 sha2 = { version = "0.10", default-features = false, features = ["oid"] }
 k256 = { version = "0.13", default-features = false, features = ["arithmetic", "ecdsa", "ecdh"] }
 chacha20poly1305 = { version = "0.10", default-features = false }
