+++ crates/microfips-core/src/noise.rs
@@ -1,2616 +1,6 @@
-//! Noise Protocol IK and XK Implementations for FIPS Interop
+//! Noise protocol re-exports from the `fips-noise` crate.
 //!
-//! Implements both handshake patterns used by FIPS:
-//! - **Noise_IK_secp256k1_ChaChaPoly_SHA256** for link-layer (FMP)
-//! - **Noise_XK_secp256k1_ChaChaPoly_SHA256** for session-layer (FSP)
-//!
-//! ## Design Note
-//!
-//! FIPS implements Noise directly (not via a spec-compliant Noise library), following
-//! only the cryptographic primitives and ordering from the Noise spec. Custom payloads
-//! (startup epoch, capability flags, negotiation) are attached to handshake messages.
-//! This is the same approach Lightning Network uses. All crypto primitives use standard
-//! Rust crates (secp256k1, chacha20poly1305, sha2, hkdf), portable to embedded targets.
-//!
-//! **Upstream (0.4.0-dev):** The `next` branch switches both link and session layers to
-//! **Noise XX** (`Noise_XX_secp256k1_ChaChaPoly_SHA256`), a 3-message handshake where
-//! neither side knows the other's static key beforehand. D2 below is eliminated by XX.
-//!
-//! ## Reference Sources
-//!
-//! - **Noise Protocol Framework** (rev 34): <https://noiseprotocol.org/noise.html>
-//!   - §4: Crypto functions — AEAD (§4.2), HASH/HKDF (§4.3)
-//!   - §5: Protocol processing — CipherState (§5.1), SymmetricState (§5.2),
-//!     HandshakeState (§5.3)
-//!   - §7.5 IK pattern: `<- s / -> e, es, s, ss / <- e, ee, se`
-//!   - §7.9 XK pattern: `<- s / -> e, es / <- e, ee, se / -> s, se`
-//!   - §13: DH functions — allows custom DH definitions
-//!   - §14: Security considerations — nonce reuse, key reuse, replay
-//! - **RFC 5869**: HMAC-based Extract-and-Expand Key Derivation Function (HKDF)
-//!   — underlying construction for `mix_key` and `Split`
-//! - **RFC 8439** (obsoletes 7539): ChaCha20 and Poly1305 for IETF Protocols
-//!   — AEAD construction. §2.8 defines the AEAD interface, §2.3 defines nonce.
-//! - **RFC 5116**: An Interface and Algorithms for Authenticated Encryption
-//! - **RFC 7748**: Elliptic Curves for Security — ECDH function
-//! - **FIPS source**: `/root/src/fips/src/noise/` on VPS (orangeclaw)
-//!   - `handshake.rs`: HandshakeState with IK/XK patterns
-//!   - `mod.rs`: CipherState with encrypt/decrypt
-//!
-//! ## FIPS Design Choices (confirmed by maintainer)
-//!
-//! These are deliberate design decisions in FIPS's custom Noise implementation,
-//! not bugs. microfips matches them for interoperability.
-//!
-//! | # | Choice | Description | Rationale |
-//! |---|--------|-------------|-----------|
-//! | D1 | Empty AAD during handshake | `AEAD_ENCRYPT(k, n, b"", plaintext)` instead of passing `h` as AAD | Custom Noise implementation with own payloads; transport keys bind via `ck`, not `h` |
-//! | D2 | IK `se` token ordering | Initiator computes `DH(e,rs)` not `DH(s,re)` | Part of custom IK implementation. Eliminated in 0.4.0-dev by switching to Noise XX. |
-//! | D3 | x-only ECDH | `SHA256(x_coordinate)` instead of raw ECDH shared secret | Required for Nostr npub compatibility (x-only keys, no parity). Same technique as BIP-340. |
-//! | D4 | Nonce format | `[0x00;4] \|\| LE64(n)` | Matches Noise spec §5.1 and RFC 8439 §2.3 — not a deviation. |
-//!
-//! ## FIPS 140-3 Compliance Gap Table
-//!
-//! | # | FIPS 140-3 Section | Gap | Recommendation |
-//! |---|-------------------|-----|----------------|
-//! | F1 | §9.9 Conditional self-tests | No KAT for ChaCha20-Poly1305 | Add RFC 8439 §2.8.2 test vectors as power-on self-test |
-//! | F2 | §9.9 Conditional self-tests | No KAT for SHA-256 | Add FIPS 180-4 Appendix B test vectors |
-//! | F3 | §9.9 Conditional self-tests | No KAT for HKDF-SHA256 | Add RFC 5869 Appendix A Test Case 1 vectors |
-//! | F4 | §9.9 Conditional self-tests | No KAT for secp256k1 ECDH | Add known base-point multiplication vectors |
-//! | F5 | SP 800-56A §5.6.2.1 | No pair-wise consistency test after keygen | Verify `DH(sk, G) == pk` after key generation |
-//! | F6 | SP 800-90B §4.1.1 | No continuous RNG test (stuck-at) | Check RNG output blocks for all-zeros |
-//! | F7 | §9.10 Power-on self-tests | No firmware integrity test | Hash code section at boot |
-//! | F8 | §8.3.2 Key zeroization | Ephemeral/session keys not zeroized on drop | Use `zeroize` crate or volatile writes |
-//! | F9 | §9.9 Error state | Self-test failure returns `Err` but no SSP halt | Implement critical error state per ISO 19790 §9.9 |
-//! | F10 | SP 800-56A Rev 3 | secp256k1 not FIPS-approved curve | Migrate to P-256 or P-384 for FIPS validation |
-//! | F11 | SP 800-38D | ChaCha20-Poly1305 not FIPS-approved AEAD | Migrate to AES-256-GCM for FIPS validation |
-//! | F12 | SP 800-56C Rev 2 | HKDF-SHA256 not FIPS-approved KDF | Use SP 800-108 or SP 800-56C compliant KDF |
-//!
-//! ## Approved Algorithm Audit
-//!
-//! | Component | Currently Used | FIPS-Approved Alternative | Spec Reference |
-//! |-----------|---------------|--------------------------|----------------|
-//! | DH | secp256k1 ECDH (custom x-only+SHA256) | ECDH P-256/P-384 (SP 800-56A Rev 3) | SP 800-56A §5.7 |
-//! | KDF | HKDF-SHA256 (RFC 5869) | SP 800-56C Rev 2 / SP 800-108 | SP 800-56C §4 |
-//! | AEAD | ChaCha20-Poly1305 (RFC 8439) | AES-256-GCM (SP 800-38D) | SP 800-38D |
-//! | Hash | SHA-256 | SHA-256 (already approved) | FIPS 180-4 |
-//! | DRBG | Hardware RNG | SP 800-90A CTR/HMAC DRBG | SP 800-90A |
-//!
-//! ## IK Handshake Pattern (Link Layer, FMP)
-//!
-//! ```text
-//!   <- s                    (pre-message: responder's static key, parity-normalized to 0x02)
-//!   -> e, es, s, ss, epoch  (msg1: 106 bytes = 33 + 49 + 24)
-//!   <- e, ee, se, epoch     (msg2: 57 bytes = 33 + 24)
-//! ```
-//!
-//! Tokens:
-//! - `e`: ephemeral public key (cleartext)
-//! - `es`: DH(e_initiator_priv, rs_responder_pub) → mix_key
-//! - `s`: static public key (encrypted)
-//! - `ss`: DH(s_initiator_priv, rs_responder_pub) → mix_key
-//! - `ee`: DH(e_initiator_priv, re_responder_pub) → mix_key
-//! - `se`: DH(e_initiator_priv, rs_responder_pub) → mix_key [D2: FIPS design choice]
-
-#[allow(deprecated)]
-use chacha20poly1305::aead::generic_array::GenericArray;
-use chacha20poly1305::aead::{AeadInPlace, KeyInit};
-use chacha20poly1305::{ChaCha20Poly1305, Tag};
-use hkdf::Hkdf;
-use k256::ecdh::diffie_hellman as raw_ecdh;
-use k256::elliptic_curve::sec1::ToEncodedPoint;
-use k256::{PublicKey, SecretKey};
-use sha2::{Digest, Sha256};
-
-use crate::identity::sha256;
-
-pub use crate::generated::fips_compat::{EPOCH_SIZE, PUBKEY_SIZE, TAG_SIZE};
-
-pub const NONCE_SIZE: usize = 12;
-
-pub const PROTOCOL_NAME: &[u8] = b"Noise_IK_secp256k1_ChaChaPoly_SHA256";
-
-pub const PROTOCOL_NAME_XK: &[u8] = b"Noise_XK_secp256k1_ChaChaPoly_SHA256";
-
-/// Protocol name for Noise XX with secp256k1 (both link and session layers).
-/// Used by FIPS `next` branch (v0.5.0+). Replaces both IK and XK.
-pub const PROTOCOL_NAME_XX: &[u8] = b"Noise_XX_secp256k1_ChaChaPoly_SHA256";
-
-/// XX handshake msg1: ephemeral only (33 bytes). No DH, no encrypted static.
-pub const XX_HANDSHAKE_MSG1_SIZE: usize = PUBKEY_SIZE;
-
-/// XX handshake msg2: ephemeral (33) + encrypted static (49) + encrypted epoch (24) = 106 bytes.
-pub const XX_HANDSHAKE_MSG2_SIZE: usize =
-    PUBKEY_SIZE + (PUBKEY_SIZE + TAG_SIZE) + (EPOCH_SIZE + TAG_SIZE);
-
-/// XX handshake msg3: encrypted static (49) + encrypted epoch (24) = 73 bytes.
-pub const XX_HANDSHAKE_MSG3_SIZE: usize = (PUBKEY_SIZE + TAG_SIZE) + (EPOCH_SIZE + TAG_SIZE);
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum NoiseError {
-    InvalidKey,
-    InvalidMessage,
-    InvalidState,
-    InvalidPublicKey,
-    DecryptionFailed,
-    EncryptionFailed,
-    BufferTooSmall,
-    MessageTooShort { expected: usize, got: usize },
-    MessageTooLarge { size: usize, max: usize },
-}
-
-/// Parity-normalize a compressed secp256k1 public key to even prefix (0x02).
-///
-/// Nostr npubs encode x-only keys without parity information. The Noise IK
-/// pre-message mixes the responder's static key into h before any messages.
-/// Both sides must mix identical bytes, so we normalize to 0x02 prefix.
-///
-/// Reference: FIPS `HandshakeState::normalize_for_premessage()` in
-/// `/root/src/fips/src/noise/handshake.rs`
-// FIPS: bd08505 noise/handshake.rs:normalize_for_premessage()
-pub fn parity_normalize(pubkey: &[u8; PUBKEY_SIZE]) -> [u8; PUBKEY_SIZE] {
-    let mut out = [0u8; PUBKEY_SIZE];
-    out[0] = 0x02;
-    out[1..].copy_from_slice(&pubkey[1..]);
-    out
-}
-
-/// x-only ECDH: compute SHA256(shared_secret_point.x_coordinate).
-///
-/// Standard ECDH returns (x, y) but we hash only the x-coordinate to make
-/// the result parity-independent. This is necessary because P and -P produce
-/// ECDH results with the same x-coordinate, so the shared secret is the same
-/// regardless of which parity the initiator assumed for the responder's key.
-///
-/// Reference: FIPS `HandshakeState::ecdh()` in
-/// `/root/src/fips/src/noise/handshake.rs` — uses `shared_secret_point()`
-/// then `SHA256(point[..32])`.
-// FIPS: bd08505 noise/handshake.rs:ecdh()
-pub fn x_only_ecdh(
-    my_secret: &[u8; 32],
-    their_pub: &[u8; PUBKEY_SIZE],
-) -> Result<[u8; 32], NoiseError> {
-    let sk = SecretKey::from_slice(my_secret).map_err(|_| NoiseError::InvalidKey)?;
-    let pk = PublicKey::from_sec1_bytes(their_pub).map_err(|_| NoiseError::InvalidKey)?;
-    let shared = raw_ecdh(sk.to_nonzero_scalar(), pk.as_affine());
-    let x = shared.raw_secret_bytes();
-    Ok(sha256(x))
-}
-
-// FIPS: bd08505 noise/mod.rs:CipherState::new()
-//
-// Compute the compressed public key for a given secret key.
-pub fn ecdh_pubkey(secret: &[u8; 32]) -> Result<[u8; PUBKEY_SIZE], NoiseError> {
-    let sk = SecretKey::from_slice(secret).map_err(|_| NoiseError::InvalidKey)?;
-    let pk = sk.public_key();
-    let encoded = pk.to_encoded_point(true);
-    let bytes = encoded.as_bytes();
-    let mut out = [0u8; PUBKEY_SIZE];
-    out.copy_from_slice(&bytes[..PUBKEY_SIZE]);
-    Ok(out)
-}
-
-/// SHA-256 of the concatenation of two byte slices: SHA256(a || b).
-///
-/// Named to match the Noise spec's `HASH(data)` primitive (§4.1).
-/// Used by `mix_hash()` to implement MixHash(data) = h = HASH(h || data).
-// FIPS: bd08505 noise/mod.rs:SymmetricState::mix_hash()
-fn hash_concat(a: &[u8], b: &[u8]) -> [u8; 32] {
-    let mut hasher = Sha256::new();
-    hasher.update(a);
-    hasher.update(b);
-    let result = hasher.finalize();
-    let mut out = [0u8; 32];
-    out.copy_from_slice(&result);
-    out
-}
-
-/// SHA-256 of a single byte slice: SHA256(data).
-///
-/// Named to match the Noise spec's `HASH(data)` primitive (§4.1).
-/// Used for protocol name hashing during handshake initialization.
-// FIPS: bd08505 noise/mod.rs:SymmetricState::initialize()
-fn hash_one(data: &[u8]) -> [u8; 32] {
-    let mut hasher = Sha256::new();
-    hasher.update(data);
-    let result = hasher.finalize();
-    let mut out = [0u8; 32];
-    out.copy_from_slice(&result);
-    out
-}
-
-/// HKDF-SHA256(key_material) → (new_chaining_key, new_cipher_key).
-///
-/// Reference: Noise spec §4.3 — HKDF with chaining_key as salt,
-/// input_key_material as IKM, zero-length info. We always request 2 outputs.
-/// FIPS uses `hkdf::Hkdf::<Sha256>::new(Some(&ck), ikm)` with expand(&[], &mut [0u8; 64]).
-// FIPS: bd08505 noise/mod.rs:SymmetricState::mix_key()
-fn mix_key(ck: &[u8; 32], ikm: &[u8; 32]) -> ([u8; 32], [u8; 32]) {
-    let hk = Hkdf::<Sha256>::new(Some(ck), ikm);
-    let mut okm = [0u8; 64];
-    hk.expand(&[], &mut okm)
-        .expect("hkdf expand 64 bytes should never fail");
-    let mut new_ck = [0u8; 32];
-    new_ck.copy_from_slice(&okm[..32]);
-    let mut k = [0u8; 32];
-    k.copy_from_slice(&okm[32..]);
-    #[cfg(feature = "std")]
-    log::trace!(
-        "mix_key: ck={:04x?}.. ikm={:04x?}.. → ck={:04x?}.. k={:04x?}..",
-        &ck[..4],
-        &ikm[..4],
-        &new_ck[..4],
-        &k[..4]
-    );
-    (new_ck, k)
-}
-
-/// Mix data into the handshake hash: h = SHA256(h || data).
-///
-/// Reference: Noise spec §5.2 SymmetricState.MixHash().
-// FIPS: bd08505 noise/mod.rs:SymmetricState::mix_hash()
-fn mix_hash(h: &[u8; 32], data: &[u8]) -> [u8; 32] {
-    hash_concat(h, data)
-}
-
-/// Construct a 12-byte nonce from a counter (4 zero bytes + 8-byte LE counter).
-///
-/// Reference: Noise spec §5.1 — "The maximum n value (2^64-1) is reserved."
-/// FIPS `CipherState::counter_to_nonce()` uses same layout: [0;4] || counter.to_le_bytes().
-// FIPS: bd08505 noise/mod.rs:CipherState::counter_to_nonce()
-fn make_nonce(n: u64) -> [u8; NONCE_SIZE] {
-    let mut nonce = [0u8; NONCE_SIZE];
-    nonce[4..].copy_from_slice(&n.to_le_bytes());
-    nonce
-}
-
-/// AEAD encrypt with ChaCha20-Poly1305.
-///
-/// `aad` is authenticated but not encrypted. During handshake, FIPS passes
-/// empty AAD (`&[]`) — see module-level docs "FIPS Deviation #1".
-///
-/// Reference: Noise spec §4.2 — ENCRYPT(k, n, ad, plaintext).
-/// Reference: FIPS `CipherState::encrypt()` in `/root/src/fips/src/noise/mod.rs`
-///   calls `cipher.encrypt(&nonce, plaintext)` with no AAD (the `Aead` trait's
-///   `encrypt` method defaults to empty AAD when called without `Payload`).
-// FIPS: bd08505 noise/mod.rs:CipherState::encrypt()
-#[allow(deprecated)]
-pub fn aead_encrypt(
-    key: &[u8; 32],
-    nonce_ctr: u64,
-    aad: &[u8],
-    plaintext: &[u8],
-    out: &mut [u8],
-) -> Result<usize, NoiseError> {
-    let cipher = ChaCha20Poly1305::new_from_slice(key).map_err(|_| NoiseError::InvalidKey)?;
-    let nonce_arr = make_nonce(nonce_ctr);
-    let nonce = GenericArray::from_slice(&nonce_arr);
-
-    let total = plaintext.len() + TAG_SIZE;
-    if out.len() < total {
-        return Err(NoiseError::BufferTooSmall);
-    }
-
-    out[..plaintext.len()].copy_from_slice(plaintext);
-    let tag = cipher
-        .encrypt_in_place_detached(nonce, aad, &mut out[..plaintext.len()])
-        .map_err(|_| NoiseError::EncryptionFailed)?;
-
-    out[plaintext.len()..total].copy_from_slice(&tag);
-    #[cfg(feature = "std")]
-    log::trace!(
-        "aead_encrypt: nonce_ctr={} pt_len={} ok",
-        nonce_ctr,
-        plaintext.len()
-    );
-    Ok(total)
-}
-
-/// AEAD decrypt with ChaCha20-Poly1305.
-///
-/// Reference: Noise spec §4.2 — DECRYPT(k, n, ad, ciphertext).
-// FIPS: bd08505 noise/mod.rs:CipherState::decrypt()
-#[allow(deprecated)]
-pub fn aead_decrypt(
-    key: &[u8; 32],
-    nonce_ctr: u64,
-    aad: &[u8],
-    ciphertext: &[u8],
-    out: &mut [u8],
-) -> Result<usize, NoiseError> {
-    if ciphertext.len() < TAG_SIZE {
-        return Err(NoiseError::InvalidMessage);
-    }
-
-    let cipher = ChaCha20Poly1305::new_from_slice(key).map_err(|_| NoiseError::InvalidKey)?;
-    let nonce_arr = make_nonce(nonce_ctr);
-    let nonce = GenericArray::from_slice(&nonce_arr);
-
-    let pt_len = ciphertext.len() - TAG_SIZE;
-    if out.len() < pt_len {
-        return Err(NoiseError::BufferTooSmall);
-    }
-
-    out[..pt_len].copy_from_slice(&ciphertext[..pt_len]);
-    let tag = Tag::from_slice(&ciphertext[pt_len..]);
-
-    cipher
-        .decrypt_in_place_detached(nonce, aad, &mut out[..pt_len], tag)
-        .map_err(|_| NoiseError::DecryptionFailed)?;
-
-    #[cfg(feature = "std")]
-    log::trace!(
-        "aead_decrypt: nonce_ctr={} ct_len={} pt_len={} ok",
-        nonce_ctr,
-        ciphertext.len(),
-        pt_len
-    );
-    Ok(pt_len)
-}
-
-// FIPS: bd08505 noise/handshake.rs:new_initiator()
-// FIPS: bd08505 noise/handshake.rs:write_message_1()
-// FIPS: bd08505 noise/handshake.rs:read_message_2()
-// FIPS: bd08505 noise/handshake.rs:SymmetricState::split()
-#[derive(Clone)]
-pub struct NoiseIkInitiator {
-    h: [u8; 32],
-    ck: [u8; 32],
-    e_priv: [u8; 32],
-    e_pub: [u8; PUBKEY_SIZE],
-    s_priv: [u8; 32],
-    rs_pub: [u8; PUBKEY_SIZE],
-    k: Option<[u8; 32]>,
-    n: u64,
-}
-
-impl NoiseIkInitiator {
-    /// Initialize the IK initiator.
-    ///
-    /// The IK pattern has a pre-message `<- s` meaning the responder's static
-    /// key is known to the initiator before the handshake begins. We
-    /// parity-normalize it and mix into h so both sides have the same hash chain.
-    ///
-    /// Reference: FIPS `HandshakeState::new_initiator()` in
-    /// `/root/src/fips/src/noise/handshake.rs` — calls
-    /// `SymmetricState::initialize(PROTOCOL_NAME_IK)` then
-    /// `mix_hash(normalize_for_premessage(&remote_static))`.
-    // FIPS: bd08505 noise/handshake.rs:new_initiator()
-    pub fn new(
-        my_ephemeral_secret: &[u8; 32],
-        my_static_secret: &[u8; 32],
-        responder_static_pub: &[u8; PUBKEY_SIZE],
-    ) -> Result<(Self, [u8; PUBKEY_SIZE]), NoiseError> {
-        let e_pub = ecdh_pubkey(my_ephemeral_secret)?;
-
-        let h = hash_one(PROTOCOL_NAME);
-        let ck = h;
-
-        let normalized_rs = parity_normalize(responder_static_pub);
-        let h = mix_hash(&h, &normalized_rs);
-
-        Ok((
-            Self {
-                h,
-                ck,
-                e_priv: *my_ephemeral_secret,
-                e_pub,
-                s_priv: *my_static_secret,
-                rs_pub: *responder_static_pub,
-                k: None,
-                n: 0,
-            },
-            e_pub,
-        ))
-    }
-
-    /// Write Noise IK message 1: `-> e, es, s, ss, epoch`
-    ///
-    /// Wire format (106 bytes):
-    /// ```text
-    ///   [e_pub: 33 bytes] [enc_s_pub: 49 bytes] [enc_epoch: 24 bytes]
-    /// ```
-    ///
-    /// Token processing order:
-    /// 1. `e`: write ephemeral public key, mix_hash(e_pub)
-    /// 2. `es`: DH(e_priv, rs_pub) → mix_key → now k is set
-    /// 3. `s`: encrypt_and_hash(s_pub) — encrypted with k from es
-    /// 4. `ss`: DH(s_priv, rs_pub) → mix_key → k changes
-    /// 5. epoch (payload): encrypt_and_hash(epoch) — encrypted with k from ss
-    ///
-    /// Reference: FIPS `HandshakeState::write_message_1()` in
-    /// `/root/src/fips/src/noise/handshake.rs`
-    // FIPS: bd08505 noise/handshake.rs:write_message_1()
-    pub fn write_message1(
-        &mut self,
-        my_static_pub: &[u8; PUBKEY_SIZE],
-        epoch: &[u8; EPOCH_SIZE],
-        out: &mut [u8],
-    ) -> Result<usize, NoiseError> {
-        let needed = PUBKEY_SIZE + (PUBKEY_SIZE + TAG_SIZE) + (EPOCH_SIZE + TAG_SIZE);
-        if out.len() < needed {
-            return Err(NoiseError::BufferTooSmall);
-        }
-
-        let mut pos = 0;
-
-        // Token: e — write ephemeral public key, mix into hash
-        out[pos..pos + PUBKEY_SIZE].copy_from_slice(&self.e_pub);
-        pos += PUBKEY_SIZE;
-        self.h = mix_hash(&self.h, &self.e_pub);
-
-        // Token: es — DH(e_initiator_priv, rs_responder_pub) → mix_key
-        let dh = x_only_ecdh(&self.e_priv, &self.rs_pub)?;
-        let (new_ck, k) = mix_key(&self.ck, &dh);
-        self.ck = new_ck;
-        self.k = Some(k);
-        self.n = 0;
-
-        // Token: s — encrypt static public key
-        // D1: empty AAD (FIPS design choice, see module docs)
-        let enc_len = aead_encrypt(
-            self.k.as_ref().unwrap(),
-            self.n,
-            &[],
-            my_static_pub,
-            &mut out[pos..],
-        )?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, &out[pos..pos + enc_len]);
-        pos += enc_len;
-
-        // Token: ss — DH(s_initiator_priv, rs_responder_pub) → mix_key
-        let ss_dh = x_only_ecdh(&self.s_priv, &self.rs_pub)?;
-        let (new_ck, k) = mix_key(&self.ck, &ss_dh);
-        self.ck = new_ck;
-        self.k = Some(k);
-        self.n = 0;
-
-        // Payload: epoch — encrypted with k from ss
-        // D1: empty AAD (FIPS design choice, see module docs)
-        let enc_len = aead_encrypt(
-            self.k.as_ref().unwrap(),
-            self.n,
-            &[], // FIPS: no AAD during handshake[],
-            epoch,
-            &mut out[pos..],
-        )?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, &out[pos..pos + enc_len]);
-        pos += enc_len;
-
-        Ok(pos)
-    }
-
-    /// Read Noise IK message 2: `<- e, ee, se, epoch`
-    ///
-    /// Wire format (57 bytes):
-    /// ```text
-    ///   [re_pub: 33 bytes] [enc_epoch: 24 bytes]
-    /// ```
-    ///
-    /// Token processing order:
-    /// 1. `e`: parse responder's ephemeral public key, mix_hash(re_pub)
-    /// 2. `ee`: DH(e_priv, re_pub) → mix_key
-    /// 3. `se`: DH(e_priv, rs_pub) → mix_key [D2: FIPS design choice (see module docs)]
-    /// 4. epoch (payload): decrypt_and_hash(enc_epoch) with k from se
-    ///
-    /// Reference: FIPS `HandshakeState::read_message_2()` in
-    /// `/root/src/fips/src/noise/handshake.rs` — note that FIPS computes
-    /// `se = DH(e_initiator_priv, rs_responder_pub)` for the initiator's
-    /// read_message_2, which swaps the key types vs the Noise spec.
-    // FIPS: bd08505 noise/handshake.rs:read_message_2()
-    pub fn read_message2(&mut self, payload: &[u8]) -> Result<[u8; EPOCH_SIZE], NoiseError> {
-        let expected = PUBKEY_SIZE + EPOCH_SIZE + TAG_SIZE;
-        if payload.len() != expected {
-            return Err(NoiseError::InvalidMessage);
-        }
-
-        let mut pos = 0;
-
-        // Token: e — parse responder's ephemeral, mix into hash
-        let re_pub: [u8; PUBKEY_SIZE] = payload[pos..pos + PUBKEY_SIZE]
-            .try_into()
-            .map_err(|_| NoiseError::InvalidMessage)?;
-        pos += PUBKEY_SIZE;
-        self.h = mix_hash(&self.h, &re_pub);
-
-        // Token: ee — DH(e_initiator_priv, re_responder_pub) → mix_key
-        let dh = x_only_ecdh(&self.e_priv, &re_pub)?;
-        let (new_ck, k) = mix_key(&self.ck, &dh);
-        self.ck = new_ck;
-        self.k = Some(k);
-        self.n = 0;
-
-        // Token: se — DH(e_initiator_priv, rs_responder_pub) → mix_key
-        // NOTE: This matches the Noise spec. The spec says initiator se=DH(s, re).
-        // DH(s, re) = DH(re, s) = DH(e_resp, s_init) which is what the responder
-        // computes. Our code computes DH(e_init, rs_resp) which equals DH(rs_resp, e_init)
-        // = DH(s_resp, e_init) = responder's es token, NOT se.
-        //
-        // This is technically wrong per the Noise spec (se ≠ es), but it matches
-        // what FIPS does. Both sides use the same computation, so it interoperates.
-        // See test `se_and_es_produce_different_keys` for details.
-        let se_dh = x_only_ecdh(&self.e_priv, &self.rs_pub)?;
-        let (new_ck, k) = mix_key(&self.ck, &se_dh);
-        self.ck = new_ck;
-        self.k = Some(k);
-        self.n = 0;
-
-        // Payload: epoch — decrypt with k from se
-        // D1: empty AAD (FIPS design choice, see module docs)
-        let enc_epoch = &payload[pos..];
-        let mut epoch_buf = [0u8; EPOCH_SIZE];
-        aead_decrypt(
-            self.k.as_ref().unwrap(),
-            self.n,
-            &[], // FIPS: no AAD during handshake[],
-            enc_epoch,
-            &mut epoch_buf,
-        )?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, enc_epoch);
-
-        Ok(epoch_buf)
-    }
-
-    /// Derive transport keys via Split().
-    ///
-    /// Reference: Noise spec §5.2 SymmetricState.Split():
-    ///   temp_k1, temp_k2 = HKDF(ck, zerolen, 2)
-    /// Returns (k_send, k_recv) for initiator-to-responder and reverse.
-    ///
-    /// Reference: FIPS `SymmetricState::split()` in
-    /// `/root/src/fips/src/noise/handshake.rs`:
-    ///   Hkdf::<Sha256>::new(Some(&self.ck), &[]) with expand(&[], &mut [0u8; 64])
-    ///   k1 = output[..32], k2 = output[32..64]
-    // FIPS: bd08505 noise/handshake.rs:SymmetricState::split()
-    pub fn finalize(&self) -> ([u8; 32], [u8; 32]) {
-        let hk = Hkdf::<Sha256>::new(Some(&self.ck), &[]);
-        let mut okm = [0u8; 64];
-        hk.expand(&[], &mut okm)
-            .expect("hkdf expand 64 bytes should never fail");
-        let mut k1 = [0u8; 32];
-        let mut k2 = [0u8; 32];
-        k1.copy_from_slice(&okm[..32]);
-        k2.copy_from_slice(&okm[32..]);
-        (k1, k2)
-    }
-}
-
-/// Noise XK Initiator for FSP session-layer handshakes.
-///
-/// Implements `Noise_XK_secp256k1_ChaChaPoly_SHA256`:
-/// ```text
-///   <- s                    (pre-message: responder's static key)
-///   -> e, es                (msg1: 33 bytes)
-///   <- e, ee, epoch          (msg2: 57 bytes)
-///   -> s, se, epoch          (msg3: 73 bytes)
-/// ```
-///
-/// The initiator knows the responder's static key upfront (from the link-layer
-/// peer index). D1: empty AAD (FIPS design choice, see module docs).
-// FIPS: bd08505 noise/handshake.rs:new_xk_initiator()
-// FIPS: bd08505 noise/handshake.rs:write_xk_message_1()
-// FIPS: bd08505 noise/handshake.rs:read_xk_message_2()
-// FIPS: bd08505 noise/handshake.rs:write_xk_message_3()
-#[derive(Clone)]
-pub struct NoiseXkInitiator {
-    h: [u8; 32],
-    ck: [u8; 32],
-    e_priv: [u8; 32],
-    e_pub: [u8; PUBKEY_SIZE],
-    s_priv: [u8; 32],
-    rs_pub: [u8; PUBKEY_SIZE],
-    re_pub: Option<[u8; PUBKEY_SIZE]>,
-    k: Option<[u8; 32]>,
-    n: u64,
-}
-
-impl NoiseXkInitiator {
-    // FIPS: bd08505 noise/handshake.rs:new_xk_initiator()
-    //
-    // Initialize the XK initiator.
-    //
-    // The XK pattern has a pre-message `<- s` meaning the responder's static
-    // key is known to the initiator before the handshake begins.
-    pub fn new(
-        my_ephemeral_secret: &[u8; 32],
-        my_static_secret: &[u8; 32],
-        responder_static_pub: &[u8; PUBKEY_SIZE],
-    ) -> Result<(Self, [u8; PUBKEY_SIZE]), NoiseError> {
-        let e_pub = ecdh_pubkey(my_ephemeral_secret)?;
-
-        let h = hash_one(PROTOCOL_NAME_XK);
-        let ck = h;
-
-        let normalized_rs = parity_normalize(responder_static_pub);
-        let h = mix_hash(&h, &normalized_rs);
-
-        Ok((
-            Self {
-                h,
-                ck,
-                e_priv: *my_ephemeral_secret,
-                e_pub,
-                s_priv: *my_static_secret,
-                rs_pub: *responder_static_pub,
-                re_pub: None,
-                k: None,
-                n: 0,
-            },
-            e_pub,
-        ))
-    }
-
-    // FIPS: bd08505 noise/handshake.rs:write_xk_message_1()
-    //
-    // Write Noise XK message 1: `-> e, es`
-    //
-    // Wire format (33 bytes):
-    // ```text
-    //   [e_pub: 33 bytes]
-    // ```
-    pub fn write_message1(&mut self, out: &mut [u8]) -> Result<usize, NoiseError> {
-        if out.len() < PUBKEY_SIZE {
-            return Err(NoiseError::BufferTooSmall);
-        }
-
-        out[..PUBKEY_SIZE].copy_from_slice(&self.e_pub);
-        self.h = mix_hash(&self.h, &self.e_pub);
-
-        let dh = x_only_ecdh(&self.e_priv, &self.rs_pub)?;
-        let (new_ck, k) = mix_key(&self.ck, &dh);
-        self.ck = new_ck;
-        self.k = Some(k);
-        self.n = 0;
-
-        Ok(PUBKEY_SIZE)
-    }
-
-    // FIPS: bd08505 noise/handshake.rs:read_xk_message_2()
-    //
-    // Read Noise XK message 2: `<- e, ee, epoch`
-    //
-    // Wire format (57 bytes):
-    // ```text
-    //   [re_pub: 33 bytes] [encrypted_epoch: 24 bytes]
-    // ```
-    //
-    // Returns the responder's epoch.
-    pub fn read_message2(&mut self, payload: &[u8]) -> Result<[u8; EPOCH_SIZE], NoiseError> {
-        if payload.len() != PUBKEY_SIZE + EPOCH_SIZE + TAG_SIZE {
-            return Err(NoiseError::InvalidMessage);
-        }
-
-        let re_pub = <[u8; PUBKEY_SIZE]>::try_from(&payload[..PUBKEY_SIZE])
-            .map_err(|_| NoiseError::InvalidMessage)?;
-        self.re_pub = Some(re_pub);
-        self.h = mix_hash(&self.h, &payload[..PUBKEY_SIZE]);
-
-        let ee = x_only_ecdh(&self.e_priv, &re_pub)?;
-        let (new_ck, k) = mix_key(&self.ck, &ee);
-        self.ck = new_ck;
-        self.k = Some(k);
-        self.n = 0;
-
-        let enc_epoch = &payload[PUBKEY_SIZE..];
-        let mut epoch_buf = [0u8; EPOCH_SIZE];
-        aead_decrypt(
-            self.k.as_ref().unwrap(),
-            self.n,
-            &[],
-            enc_epoch,
-            &mut epoch_buf,
-        )?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, enc_epoch);
-
-        Ok(epoch_buf)
-    }
-
-    // FIPS: bd08505 noise/handshake.rs:write_xk_message_3()
-    //
-    // Write Noise XK message 3: `-> s, se, epoch`
-    //
-    // Wire format (73 bytes):
-    // ```text
-    //   [encrypted_static: 49 bytes] [encrypted_epoch: 24 bytes]
-    // ```
-    pub fn write_message3(
-        &mut self,
-        my_static_pub: &[u8; PUBKEY_SIZE],
-        epoch: &[u8; EPOCH_SIZE],
-        out: &mut [u8],
-    ) -> Result<usize, NoiseError> {
-        let needed = (PUBKEY_SIZE + TAG_SIZE) + (EPOCH_SIZE + TAG_SIZE);
-        if out.len() < needed {
-            return Err(NoiseError::BufferTooSmall);
-        }
-
-        let mut pos = 0;
-
-        let enc_len = aead_encrypt(
-            self.k.as_ref().unwrap(),
-            self.n,
-            &[], // FIPS: no AAD during handshake[],
-            my_static_pub,
-            &mut out[pos..],
-        )?;
-        pos += enc_len;
-        self.h = mix_hash(&self.h, &out[..enc_len]);
-        self.n += 1;
-
-        // Token: se — DH(s_initiator, re_responder) → mix_key
-        let re_pub = self.re_pub.unwrap();
-        let se = x_only_ecdh(&self.s_priv, &re_pub)?;
-        let (new_ck, k) = mix_key(&self.ck, &se);
-        self.ck = new_ck;
-        self.k = Some(k);
-        self.n = 0;
-
-        let enc_epoch_len = aead_encrypt(
-            self.k.as_ref().unwrap(),
-            self.n,
-            &[], // FIPS: no AAD during handshake[],
-            epoch,
-            &mut out[pos..],
-        )?;
-        pos += enc_epoch_len;
-        self.h = mix_hash(&self.h, &out[pos - enc_epoch_len..pos]);
-        self.n += 1;
-
-        Ok(pos)
-    }
-
-    // FIPS: bd08505 noise/handshake.rs:SymmetricState::split()
-    // FIPS: bd08505 noise/handshake.rs:into_session()
-    //
-    // Derive transport keys via Split() (same as IK).
-    pub fn finalize(&self) -> ([u8; 32], [u8; 32]) {
-        let hk = Hkdf::<Sha256>::new(Some(&self.ck), &[]);
-        let mut okm = [0u8; 64];
-        hk.expand(&[], &mut okm)
-            .expect("hkdf expand 64 bytes should never fail");
-        let mut k1 = [0u8; 32];
-        let mut k2 = [0u8; 32];
-        k1.copy_from_slice(&okm[..32]);
-        k2.copy_from_slice(&okm[32..]);
-        (k1, k2)
-    }
-}
-
-// FIPS: bd08505 noise/handshake.rs:new_xk_responder()
-// FIPS: bd08505 noise/handshake.rs:write_xk_message_2()
-// FIPS: bd08505 noise/handshake.rs:read_xk_message_3()
-//
-// Noise XK Responder for FSP session-layer handshakes.
-//
-// Implements `Noise_XK_secp256k1_ChaChaPoly_SHA256` responder side:
-// ```text
-//   <- s                    (pre-message: our static key)
-//   -> e, es                (msg1: 33 bytes — read by new())
-//   <- e, ee, epoch          (msg2: 57 bytes — write_message2())
-//   -> s, se, epoch          (msg3: 73 bytes — read_message3())
-// ```
-pub struct NoiseXkResponder {
-    h: [u8; 32],
-    ck: [u8; 32],
-    ei_pub: [u8; PUBKEY_SIZE],
-    e_priv: Option<[u8; 32]>,
-    k: Option<[u8; 32]>,
-    n: u64,
-}
-
-impl NoiseXkResponder {
-    // FIPS: bd08505 noise/handshake.rs:new_xk_responder()
-    pub fn new(
-        responder_static_secret: &[u8; 32],
-        initiator_ephemeral_pub: &[u8; PUBKEY_SIZE],
-    ) -> Result<Self, NoiseError> {
-        let h = hash_one(PROTOCOL_NAME_XK);
-        let ck = h;
-
-        let normalized_s = parity_normalize(&ecdh_pubkey(responder_static_secret)?);
-        let h = mix_hash(&h, &normalized_s);
-
-        let h = mix_hash(&h, initiator_ephemeral_pub);
-
-        let es = x_only_ecdh(responder_static_secret, initiator_ephemeral_pub)?;
-        let (ck, k) = mix_key(&ck, &es);
-
-        Ok(Self {
-            h,
-            ck,
-            ei_pub: *initiator_ephemeral_pub,
-            e_priv: None,
-            k: Some(k),
-            n: 0,
-        })
-    }
-
-    // FIPS: bd08505 noise/handshake.rs:write_xk_message_2()
-    //
-    // Write Noise XK message 2: `<- e, ee, epoch`
-    //
-    // Wire format (57 bytes):
-    // ```text
-    //   [e_pub: 33 bytes] [encrypted_epoch: 24 bytes]
-    // ```
-    pub fn write_message2(
-        &mut self,
-        responder_ephemeral_secret: &[u8; 32],
-        epoch: &[u8; EPOCH_SIZE],
-        out: &mut [u8],
-    ) -> Result<usize, NoiseError> {
-        let needed = PUBKEY_SIZE + EPOCH_SIZE + TAG_SIZE;
-        if out.len() < needed {
-            return Err(NoiseError::BufferTooSmall);
-        }
-        self.e_priv = Some(*responder_ephemeral_secret);
-        let e_pub = ecdh_pubkey(responder_ephemeral_secret)?;
-        let mut pos = 0;
-
-        out[pos..pos + PUBKEY_SIZE].copy_from_slice(&e_pub);
-        pos += PUBKEY_SIZE;
-        self.h = mix_hash(&self.h, &e_pub);
-
-        let ee = x_only_ecdh(responder_ephemeral_secret, &self.ei_pub)?;
-        let (new_ck, k) = mix_key(&self.ck, &ee);
-        self.ck = new_ck;
-        self.k = Some(k);
-        self.n = 0;
-
-        let enc_len = aead_encrypt(
-            self.k.as_ref().unwrap(),
-            self.n,
-            &[], // FIPS: no AAD during handshake[],
-            epoch,
-            &mut out[pos..],
-        )?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, &out[pos..pos + enc_len]);
-        pos += enc_len;
-
-        Ok(pos)
-    }
-
-    // FIPS: bd08505 noise/handshake.rs:read_xk_message_3()
-    //
-    // Read Noise XK message 3: `-> s, se, epoch`
-    //
-    // Wire format (73 bytes):
-    // ```text
-    //   [encrypted_static: 49 bytes] [encrypted_epoch: 24 bytes]
-    // ```
-    //
-    // Returns (initiator_static_pub, initiator_epoch).
-    pub fn read_message3(
-        &mut self,
-        payload: &[u8],
-    ) -> Result<([u8; PUBKEY_SIZE], [u8; EPOCH_SIZE]), NoiseError> {
-        let expected = (PUBKEY_SIZE + TAG_SIZE) + (EPOCH_SIZE + TAG_SIZE);
-        if payload.len() != expected {
-            return Err(NoiseError::InvalidMessage);
-        }
-
-        let enc_static = &payload[..PUBKEY_SIZE + TAG_SIZE];
-        let mut static_buf = [0u8; PUBKEY_SIZE];
-        aead_decrypt(
-            self.k.as_ref().unwrap(),
-            self.n,
-            &[], // FIPS: no AAD during handshake[],
-            enc_static,
-            &mut static_buf,
-        )?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, enc_static);
-
-        let se = x_only_ecdh(
-            self.e_priv.as_ref().ok_or(NoiseError::InvalidState)?,
-            &static_buf,
-        )?;
-        let (new_ck, k) = mix_key(&self.ck, &se);
-        self.ck = new_ck;
-        self.k = Some(k);
-        self.n = 0;
-
-        let enc_epoch = &payload[PUBKEY_SIZE + TAG_SIZE..];
-        let mut epoch_buf = [0u8; EPOCH_SIZE];
-        aead_decrypt(
-            self.k.as_ref().unwrap(),
-            self.n,
-            &[], // FIPS: no AAD during handshake[],
-            enc_epoch,
-            &mut epoch_buf,
-        )?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, enc_epoch);
-
-        Ok((static_buf, epoch_buf))
-    }
-
-    // FIPS: bd08505 noise/handshake.rs:SymmetricState::split()
-    // FIPS: bd08505 noise/handshake.rs:into_session()
-    //
-    // Derive transport keys via Split().
-    // Returns (k_recv, k_send) for responder.
-    // k_recv = initiator→responder, k_send = responder→initiator.
-    pub fn finalize(&self) -> ([u8; 32], [u8; 32]) {
-        let hk = Hkdf::<Sha256>::new(Some(&self.ck), &[]);
-        let mut okm = [0u8; 64];
-        hk.expand(&[], &mut okm)
-            .expect("hkdf expand 64 bytes should never fail");
-        let mut k1 = [0u8; 32];
-        let mut k2 = [0u8; 32];
-        k1.copy_from_slice(&okm[..32]);
-        k2.copy_from_slice(&okm[32..]);
-        (k1, k2)
-    }
-}
-
-// ===========================================================================
-// Noise XX — Link + Session Layer (FIPS next branch, v0.5.0+)
-// ===========================================================================
-//
-// Replaces both IK (link) and XK (session) with a single unified pattern.
-// Neither side knows the other's static key before the handshake.
-//
-// XX Handshake Pattern:
-//   -> e                        (msg1: 33B — ephemeral only, no DH)
-//   <- e, ee, s, es, epoch      (msg2: 106B — ephemeral + encrypted static + epoch)
-//   -> s, se, epoch             (msg3: 73B — encrypted static + epoch)
-//
-// Identity timing:
-//   msg1: no identity
-//   msg2: responder reveals static to initiator
-//   msg3: initiator reveals static to responder
-
-/// Noise XX Initiator for both link-layer (FMP) and session-layer (FSP).
-#[derive(Clone, Copy)]
-pub struct NoiseXxInitiator {
-    h: [u8; 32],
-    ck: [u8; 32],
-    e_priv: [u8; 32],
-    e_pub: [u8; PUBKEY_SIZE],
-    s_priv: [u8; 32],
-    re_pub: Option<[u8; PUBKEY_SIZE]>,
-    rs_pub: Option<[u8; PUBKEY_SIZE]>,
-    k: Option<[u8; 32]>,
-    n: u64,
-}
-
-impl NoiseXxInitiator {
-    pub fn new(
-        my_ephemeral_secret: &[u8; 32],
-        my_static_secret: &[u8; 32],
-    ) -> Result<(Self, [u8; PUBKEY_SIZE]), NoiseError> {
-        let e_pub = ecdh_pubkey(my_ephemeral_secret)?;
-        let h = hash_one(PROTOCOL_NAME_XX);
-        Ok((
-            Self {
-                h,
-                ck: h,
-                e_priv: *my_ephemeral_secret,
-                e_pub,
-                s_priv: *my_static_secret,
-                re_pub: None,
-                rs_pub: None,
-                k: None,
-                n: 0,
-            },
-            e_pub,
-        ))
-    }
-
-    /// Write XX msg1: `-> e` (33 bytes). Ephemeral key only, no DH.
-    pub fn write_message1(&mut self, out: &mut [u8]) -> Result<usize, NoiseError> {
-        if out.len() < XX_HANDSHAKE_MSG1_SIZE {
-            return Err(NoiseError::BufferTooSmall);
-        }
-        out[..PUBKEY_SIZE].copy_from_slice(&self.e_pub);
-        self.h = mix_hash(&self.h, &self.e_pub);
-        Ok(PUBKEY_SIZE)
-    }
-
-    /// Read XX msg2: `<- e, ee, s, es, epoch` (106 bytes).
-    /// Returns (responder_static_pub, responder_epoch).
-    pub fn read_message2(
-        &mut self,
-        payload: &[u8],
-    ) -> Result<([u8; PUBKEY_SIZE], [u8; EPOCH_SIZE]), NoiseError> {
-        if payload.len() != XX_HANDSHAKE_MSG2_SIZE {
-            return Err(NoiseError::InvalidMessage);
-        }
-
-        let re_pub: [u8; PUBKEY_SIZE] = payload[..PUBKEY_SIZE].try_into().unwrap();
-        self.re_pub = Some(re_pub);
-        self.h = mix_hash(&self.h, &re_pub);
-
-        let ee = x_only_ecdh(&self.e_priv, &re_pub)?;
-        let (ck, k) = mix_key(&self.ck, &ee);
-        self.ck = ck;
-        self.k = Some(k);
-        self.n = 0;
-
-        let enc_rs = &payload[PUBKEY_SIZE..PUBKEY_SIZE + PUBKEY_SIZE + TAG_SIZE];
-        let mut rs_pub = [0u8; PUBKEY_SIZE];
-        aead_decrypt(self.k.as_ref().unwrap(), self.n, &[], enc_rs, &mut rs_pub)?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, enc_rs);
-        self.rs_pub = Some(rs_pub);
-
-        let es = x_only_ecdh(&self.e_priv, &rs_pub)?;
-        let (ck, k) = mix_key(&self.ck, &es);
-        self.ck = ck;
-        self.k = Some(k);
-        self.n = 0;
-
-        let enc_epoch = &payload[PUBKEY_SIZE + PUBKEY_SIZE + TAG_SIZE..];
-        let mut epoch_buf = [0u8; EPOCH_SIZE];
-        aead_decrypt(
-            self.k.as_ref().unwrap(),
-            self.n,
-            &[],
-            enc_epoch,
-            &mut epoch_buf,
-        )?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, enc_epoch);
-
-        Ok((rs_pub, epoch_buf))
-    }
-
-    /// Write XX msg3: `-> s, se, epoch` (73 bytes).
-    pub fn write_message3(
-        &mut self,
-        my_static_pub: &[u8; PUBKEY_SIZE],
-        epoch: &[u8; EPOCH_SIZE],
-        out: &mut [u8],
-    ) -> Result<usize, NoiseError> {
-        if out.len() < XX_HANDSHAKE_MSG3_SIZE {
-            return Err(NoiseError::BufferTooSmall);
-        }
-        let re_pub = self.re_pub.ok_or(NoiseError::InvalidState)?;
-        let mut pos = 0;
-
-        let enc_len = aead_encrypt(
-            self.k.as_ref().unwrap(),
-            self.n,
-            &[],
-            my_static_pub,
-            &mut out[pos..],
-        )?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, &out[pos..pos + enc_len]);
-        pos += enc_len;
-
-        let se = x_only_ecdh(&self.s_priv, &re_pub)?;
-        let (ck, k) = mix_key(&self.ck, &se);
-        self.ck = ck;
-        self.k = Some(k);
-        self.n = 0;
-
-        let enc_len = aead_encrypt(
-            self.k.as_ref().unwrap(),
-            self.n,
-            &[],
-            epoch,
-            &mut out[pos..],
-        )?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, &out[pos..pos + enc_len]);
-        pos += enc_len;
-
-        Ok(pos)
-    }
-
-    /// Encrypt negotiation payload after msg3, before finalize.
-    pub fn encrypt_payload(
-        &mut self,
-        plaintext: &[u8],
-        out: &mut [u8],
-    ) -> Result<usize, NoiseError> {
-        let enc_len = aead_encrypt(self.k.as_ref().unwrap(), self.n, &[], plaintext, out)?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, &out[..enc_len]);
-        Ok(enc_len)
-    }
-
-    /// Decrypt negotiation payload after msg2, before finalize.
-    pub fn decrypt_payload(
-        &mut self,
-        ciphertext: &[u8],
-        out: &mut [u8],
-    ) -> Result<usize, NoiseError> {
-        let dec_len = aead_decrypt(self.k.as_ref().unwrap(), self.n, &[], ciphertext, out)?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, ciphertext);
-        Ok(dec_len)
-    }
-
-    /// Derive transport keys via Split(). Returns (k1, k2) = (c1, c2).
-    /// c1 = initiator→responder, c2 = responder→initiator.
-    pub fn finalize(&self) -> ([u8; 32], [u8; 32]) {
-        let hk = Hkdf::<Sha256>::new(Some(&self.ck), &[]);
-        let mut okm = [0u8; 64];
-        hk.expand(&[], &mut okm)
-            .expect("hkdf expand 64 bytes should never fail");
-        let mut k1 = [0u8; 32];
-        let mut k2 = [0u8; 32];
-        k1.copy_from_slice(&okm[..32]);
-        k2.copy_from_slice(&okm[32..]);
-        (k1, k2)
-    }
-
-    pub fn remote_static(&self) -> Option<&[u8; PUBKEY_SIZE]> {
-        self.rs_pub.as_ref()
-    }
-}
-
-/// Noise XX Responder for both link-layer (FMP) and session-layer (FSP).
-#[derive(Clone, Copy)]
-pub struct NoiseXxResponder {
-    h: [u8; 32],
-    ck: [u8; 32],
-    s_priv: [u8; 32],
-    s_pub: [u8; PUBKEY_SIZE],
-    e_priv: Option<[u8; 32]>,
-    e_pub: Option<[u8; PUBKEY_SIZE]>,
-    ei_pub: Option<[u8; PUBKEY_SIZE]>,
-    rs_pub: Option<[u8; PUBKEY_SIZE]>,
-    k: Option<[u8; 32]>,
-    n: u64,
-}
-
-impl NoiseXxResponder {
-    pub fn new(my_static_secret: &[u8; 32]) -> Result<Self, NoiseError> {
-        let s_pub = ecdh_pubkey(my_static_secret)?;
-        let h = hash_one(PROTOCOL_NAME_XX);
-        Ok(Self {
-            h,
-            ck: h,
-            s_priv: *my_static_secret,
-            s_pub,
-            e_priv: None,
-            e_pub: None,
-            ei_pub: None,
-            rs_pub: None,
-            k: None,
-            n: 0,
-        })
-    }
-
-    /// Read XX msg1: `-> e` (33 bytes). Parse ephemeral, no DH.
-    pub fn read_message1(&mut self, payload: &[u8]) -> Result<(), NoiseError> {
-        if payload.len() != XX_HANDSHAKE_MSG1_SIZE {
-            return Err(NoiseError::InvalidMessage);
-        }
-        let ei_pub: [u8; PUBKEY_SIZE] = payload[..PUBKEY_SIZE].try_into().unwrap();
-        self.ei_pub = Some(ei_pub);
-        self.h = mix_hash(&self.h, &ei_pub);
-        Ok(())
-    }
-
-    /// Write XX msg2: `<- e, ee, s, es, epoch` (106 bytes).
-    pub fn write_message2(
-        &mut self,
-        my_ephemeral_secret: &[u8; 32],
-        epoch: &[u8; EPOCH_SIZE],
-        out: &mut [u8],
-    ) -> Result<usize, NoiseError> {
-        if out.len() < XX_HANDSHAKE_MSG2_SIZE {
-            return Err(NoiseError::BufferTooSmall);
-        }
-        let ei_pub = self.ei_pub.ok_or(NoiseError::InvalidState)?;
-
-        self.e_priv = Some(*my_ephemeral_secret);
-        let e_pub = ecdh_pubkey(my_ephemeral_secret)?;
-        self.e_pub = Some(e_pub);
-        let mut pos = 0;
-
-        out[..PUBKEY_SIZE].copy_from_slice(&e_pub);
-        self.h = mix_hash(&self.h, &e_pub);
-        pos += PUBKEY_SIZE;
-
-        let ee = x_only_ecdh(my_ephemeral_secret, &ei_pub)?;
-        let (ck, k) = mix_key(&self.ck, &ee);
-        self.ck = ck;
-        self.k = Some(k);
-        self.n = 0;
-
-        let enc_len = aead_encrypt(
-            self.k.as_ref().unwrap(),
-            self.n,
-            &[],
-            &self.s_pub,
-            &mut out[pos..],
-        )?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, &out[pos..pos + enc_len]);
-        pos += enc_len;
-
-        let es = x_only_ecdh(&self.s_priv, &ei_pub)?;
-        let (ck, k) = mix_key(&self.ck, &es);
-        self.ck = ck;
-        self.k = Some(k);
-        self.n = 0;
-
-        let enc_len = aead_encrypt(
-            self.k.as_ref().unwrap(),
-            self.n,
-            &[],
-            epoch,
-            &mut out[pos..],
-        )?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, &out[pos..pos + enc_len]);
-        pos += enc_len;
-
-        Ok(pos)
-    }
-
-    /// Read XX msg3: `-> s, se, epoch` (73 bytes).
-    /// Returns (initiator_static_pub, initiator_epoch).
-    pub fn read_message3(
-        &mut self,
-        payload: &[u8],
-    ) -> Result<([u8; PUBKEY_SIZE], [u8; EPOCH_SIZE]), NoiseError> {
-        if payload.len() != XX_HANDSHAKE_MSG3_SIZE {
-            return Err(NoiseError::InvalidMessage);
-        }
-
-        let enc_rs = &payload[..PUBKEY_SIZE + TAG_SIZE];
-        let mut rs_pub = [0u8; PUBKEY_SIZE];
-        aead_decrypt(self.k.as_ref().unwrap(), self.n, &[], enc_rs, &mut rs_pub)?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, enc_rs);
-        self.rs_pub = Some(rs_pub);
-
-        let se = x_only_ecdh(self.e_priv.as_ref().unwrap(), &rs_pub)?;
-        let (ck, k) = mix_key(&self.ck, &se);
-        self.ck = ck;
-        self.k = Some(k);
-        self.n = 0;
-
-        let enc_epoch = &payload[PUBKEY_SIZE + TAG_SIZE..];
-        let mut epoch_buf = [0u8; EPOCH_SIZE];
-        aead_decrypt(
-            self.k.as_ref().unwrap(),
-            self.n,
-            &[],
-            enc_epoch,
-            &mut epoch_buf,
-        )?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, enc_epoch);
-
-        Ok((rs_pub, epoch_buf))
-    }
-
-    /// Encrypt negotiation payload after msg2, before finalize.
-    pub fn encrypt_payload(
-        &mut self,
-        plaintext: &[u8],
-        out: &mut [u8],
-    ) -> Result<usize, NoiseError> {
-        let enc_len = aead_encrypt(self.k.as_ref().unwrap(), self.n, &[], plaintext, out)?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, &out[..enc_len]);
-        Ok(enc_len)
-    }
-
-    /// Decrypt negotiation payload after msg3, before finalize.
-    pub fn decrypt_payload(
-        &mut self,
-        ciphertext: &[u8],
-        out: &mut [u8],
-    ) -> Result<usize, NoiseError> {
-        let dec_len = aead_decrypt(self.k.as_ref().unwrap(), self.n, &[], ciphertext, out)?;
-        self.n += 1;
-        self.h = mix_hash(&self.h, ciphertext);
-        Ok(dec_len)
-    }
-
-    /// Derive transport keys via Split(). Returns (k1, k2) = (c1, c2).
-    /// c1 = initiator→responder, c2 = responder→initiator.
-    pub fn finalize(&self) -> ([u8; 32], [u8; 32]) {
-        let hk = Hkdf::<Sha256>::new(Some(&self.ck), &[]);
-        let mut okm = [0u8; 64];
-        hk.expand(&[], &mut okm)
-            .expect("hkdf expand 64 bytes should never fail");
-        let mut k1 = [0u8; 32];
-        let mut k2 = [0u8; 32];
-        k1.copy_from_slice(&okm[..32]);
-        k2.copy_from_slice(&okm[32..]);
-        (k1, k2)
-    }
-
-    pub fn remote_static(&self) -> Option<&[u8; PUBKEY_SIZE]> {
-        self.rs_pub.as_ref()
-    }
-}
-
-/// Generate a fresh random secp256k1 keypair for testing.
-///
-/// Uses OS-level randomness so each test invocation exercises the protocol
-/// with different keys, proving correctness is not tied to one specific keypair.
-#[cfg(test)]
-fn test_keypair() -> ([u8; 32], [u8; PUBKEY_SIZE]) {
-    use k256::SecretKey;
-    use rand::RngCore;
-
-    let mut rng = rand::rng();
-    let mut secret = [0u8; 32];
-    loop {
-        rng.fill_bytes(&mut secret);
-        if SecretKey::from_slice(&secret).is_ok() {
-            break;
-        }
-    }
-    let pub_key = ecdh_pubkey(&secret).unwrap();
-    (secret, pub_key)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn parity_normalize_forces_even_prefix() {
-        let mut odd_key = [0u8; 33];
-        odd_key[0] = 0x03;
-        odd_key[1] = 0xAB;
-        let normalized = parity_normalize(&odd_key);
-        assert_eq!(normalized[0], 0x02);
-        assert_eq!(normalized[1], 0xAB);
-    }
-
-    #[test]
-    fn parity_normalize_preserves_even() {
-        let mut even_key = [0u8; 33];
-        even_key[0] = 0x02;
-        even_key[1] = 0xCD;
-        let normalized = parity_normalize(&even_key);
-        assert_eq!(normalized[0], 0x02);
-        assert_eq!(normalized[1], 0xCD);
-    }
-
-    #[test]
-    fn ecdh_keypair_roundtrip() {
-        let (secret, pub_key) = test_keypair();
-        assert!(
-            pub_key[0] == 0x02 || pub_key[0] == 0x03,
-            "valid compressed prefix"
-        );
-        let recomputed = ecdh_pubkey(&secret).unwrap();
-        assert_eq!(pub_key, recomputed);
-    }
-
-    #[test]
-    fn x_only_ecdh_is_deterministic() {
-        let (secret_a, pub_a) = test_keypair();
-        let secret_b: [u8; 32] = [
-            0xAA, 0xBB, 0xCC, 0xDD, 0xEE, 0xFF, 0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77,
-            0x88, 0x99, 0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, 0x88, 0x99, 0xAA, 0xBB,
-            0xCC, 0xDD, 0xEE, 0xFF,
-        ];
-        let pub_b = ecdh_pubkey(&secret_b).unwrap();
-
-        let dh1 = x_only_ecdh(&secret_a, &pub_b).unwrap();
-        let dh2 = x_only_ecdh(&secret_b, &pub_a).unwrap();
-        assert_eq!(dh1, dh2);
-    }
-
-    #[test]
-    fn aead_encrypt_decrypt_roundtrip() {
-        let key = [0x42u8; 32];
-        let plaintext = b"hello noise";
-        let aad = b"associated";
-
-        let mut ciphertext = [0u8; 256];
-        let ct_len = aead_encrypt(&key, 0, aad, plaintext, &mut ciphertext).unwrap();
-
-        let mut decrypted = [0u8; 256];
-        let pt_len = aead_decrypt(&key, 0, aad, &ciphertext[..ct_len], &mut decrypted).unwrap();
-
-        assert_eq!(&decrypted[..pt_len], plaintext);
-    }
-
-    #[test]
-    fn aead_wrong_key_fails() {
-        let key = [0x42u8; 32];
-        let wrong_key = [0x43u8; 32];
-        let plaintext = b"hello noise";
-
-        let mut ciphertext = [0u8; 256];
-        let ct_len = aead_encrypt(&key, 0, b"", plaintext, &mut ciphertext).unwrap();
-
-        let mut decrypted = [0u8; 256];
-        let result = aead_decrypt(&wrong_key, 0, b"", &ciphertext[..ct_len], &mut decrypted);
-        assert_eq!(result, Err(NoiseError::DecryptionFailed));
-    }
-
-    #[test]
-    fn aead_wrong_nonce_fails() {
-        let key = [0x42u8; 32];
-        let plaintext = b"hello noise";
-
-        let mut ciphertext = [0u8; 256];
-        let ct_len = aead_encrypt(&key, 0, b"", plaintext, &mut ciphertext).unwrap();
-
-        let mut decrypted = [0u8; 256];
-        let result = aead_decrypt(&key, 1, b"", &ciphertext[..ct_len], &mut decrypted);
-        assert_eq!(result, Err(NoiseError::DecryptionFailed));
-    }
-
-    #[test]
-    fn aead_wrong_aad_fails() {
-        let key = [0x42u8; 32];
-        let plaintext = b"hello noise";
-
-        let mut ciphertext = [0u8; 256];
-        let ct_len = aead_encrypt(&key, 0, b"correct_aad", plaintext, &mut ciphertext).unwrap();
-
-        let mut decrypted = [0u8; 256];
-        let result = aead_decrypt(&key, 0, b"wrong_aad", &ciphertext[..ct_len], &mut decrypted);
-        assert_eq!(result, Err(NoiseError::DecryptionFailed));
-    }
-
-    #[test]
-    fn mix_key_deterministic() {
-        let ck = [0x01u8; 32];
-        let ikm = [0x02u8; 32];
-        let (ck1, k1) = mix_key(&ck, &ikm);
-        let (ck2, k2) = mix_key(&ck, &ikm);
-        assert_eq!(ck1, ck2);
-        assert_eq!(k1, k2);
-        assert_ne!(ck1, ck);
-        assert_ne!(k1, ck);
-    }
-
-    #[test]
-    fn noise_ik_initiator_creates_state() {
-        let (eph_secret, _) = test_keypair();
-        let (s_secret, _) = test_keypair();
-        let responder_pub = [0x02u8; 33];
-        let (state, e_pub) = NoiseIkInitiator::new(&eph_secret, &s_secret, &responder_pub).unwrap();
-        assert_eq!(e_pub, ecdh_pubkey(&eph_secret).unwrap());
-        assert_eq!(state.n, 0);
-    }
-
-    #[test]
-    fn noise_ik_msg1_size() {
-        let (eph_secret, _) = test_keypair();
-        let (s_secret, _) = test_keypair();
-        let responder_pub = [0x02u8; 33];
-        let (mut state, _) = NoiseIkInitiator::new(&eph_secret, &s_secret, &responder_pub).unwrap();
-
-        let my_static = ecdh_pubkey(&[0xAA; 32]).unwrap();
-        let epoch = [0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08];
-
-        let mut out = [0u8; 256];
-        let msg_len = state.write_message1(&my_static, &epoch, &mut out).unwrap();
-
-        // 33 (e_pub) + 49 (enc_s_pub = 33 + 16 tag) + 24 (enc_epoch = 8 + 16 tag) = 106
-        assert_eq!(msg_len, 106);
-    }
-
-    #[test]
-    fn noise_ik_msg1_contains_ephemeral_pubkey() {
-        let (eph_secret, _) = test_keypair();
-        let (s_secret, _) = test_keypair();
-        let responder_pub = [0x02u8; 33];
-        let (mut state, e_pub) =
-            NoiseIkInitiator::new(&eph_secret, &s_secret, &responder_pub).unwrap();
-
-        let my_static = ecdh_pubkey(&[0xAA; 32]).unwrap();
-        let epoch = [0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08];
-
-        let mut out = [0u8; 256];
-        state.write_message1(&my_static, &epoch, &mut out).unwrap();
-
-        assert_eq!(&out[..33], &e_pub);
-    }
-
-    #[test]
-    fn noise_ik_msg1_enc_static_is_correct_size() {
-        let (eph_secret, _) = test_keypair();
-        let (s_secret, _) = test_keypair();
-        let responder_pub = [0x02u8; 33];
-        let (mut state, _) = NoiseIkInitiator::new(&eph_secret, &s_secret, &responder_pub).unwrap();
-
-        let my_static = ecdh_pubkey(&[0xAA; 32]).unwrap();
-        let epoch = [0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08];
-
-        let mut out = [0u8; 256];
-        state.write_message1(&my_static, &epoch, &mut out).unwrap();
-
-        let enc_static = &out[33..33 + 49];
-        assert_eq!(enc_static.len(), 49);
-    }
-
-    #[test]
-    fn noise_ik_msg1_enc_epoch_is_correct_size() {
-        let (eph_secret, _) = test_keypair();
-        let (s_secret, _) = test_keypair();
-        let responder_pub = [0x02u8; 33];
-        let (mut state, _) = NoiseIkInitiator::new(&eph_secret, &s_secret, &responder_pub).unwrap();
-
-        let my_static = ecdh_pubkey(&[0xAA; 32]).unwrap();
-        let epoch = [0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08];
-
-        let mut out = [0u8; 256];
-        state.write_message1(&my_static, &epoch, &mut out).unwrap();
-
-        let enc_epoch = &out[82..106];
-        assert_eq!(enc_epoch.len(), 24);
-    }
-
-    #[test]
-    fn protocol_name_hash() {
-        let h = hash_one(PROTOCOL_NAME);
-        assert_ne!(h, [0u8; 32]);
-        assert_eq!(h, sha256(PROTOCOL_NAME));
-    }
-}
-
-mod responder_pub {
-    use super::*;
-
-    pub struct NoiseIkResponder {
-        h: [u8; 32],
-        ck: [u8; 32],
-        s_priv: [u8; 32],
-        ei_pub: [u8; PUBKEY_SIZE],
-        rs_pub: Option<[u8; PUBKEY_SIZE]>,
-        k: Option<[u8; 32]>,
-        n: u64,
-    }
-
-    impl NoiseIkResponder {
-        // FIPS: bd08505 noise/handshake.rs:new_responder()
-        //
-        // Create a new IK responder.
-        //
-        // Returns `Err` if the responder's static key or the initiator's
-        // ephemeral public key is invalid (not on the secp256k1 curve).
-        pub fn new(
-            responder_static_secret: &[u8; 32],
-            initiator_ephemeral_pub: &[u8; PUBKEY_SIZE],
-        ) -> Result<Self, NoiseError> {
-            let h = hash_one(PROTOCOL_NAME);
-            let ck = h;
-            let normalized_rs = parity_normalize(
-                &ecdh_pubkey(responder_static_secret).map_err(|_| NoiseError::InvalidPublicKey)?,
-            );
-            let h = mix_hash(&h, &normalized_rs);
-            let h = mix_hash(&h, initiator_ephemeral_pub);
-
-            let dh = x_only_ecdh(responder_static_secret, initiator_ephemeral_pub)
-                .map_err(|_| NoiseError::DecryptionFailed)?;
-            let (ck, k) = mix_key(&ck, &dh);
-
-            Ok(Self {
-                h,
-                ck,
-                s_priv: *responder_static_secret,
-                ei_pub: *initiator_ephemeral_pub,
-                rs_pub: None,
-                k: Some(k),
-                n: 0,
-            })
-        }
-
-        // FIPS: bd08505 noise/handshake.rs:read_message_1()
-        //
-        // Read message 1 from the initiator.
-        //
-        // Returns `(initiator_static_pub, initiator_epoch)`.
-        // Returns `Err` if payload is too short or AEAD decryption fails.
-        pub fn read_message1(
-            &mut self,
-            payload: &[u8],
-        ) -> Result<([u8; PUBKEY_SIZE], [u8; EPOCH_SIZE]), NoiseError> {
-            const MIN_MSG1_PAYLOAD: usize = PUBKEY_SIZE + TAG_SIZE + EPOCH_SIZE + TAG_SIZE;
-            if payload.len() < MIN_MSG1_PAYLOAD {
-                return Err(NoiseError::MessageTooShort {
-                    expected: MIN_MSG1_PAYLOAD,
-                    got: payload.len(),
-                });
-            }
-
-            let enc_static = &payload[..49];
-            let mut static_buf = [0u8; PUBKEY_SIZE];
-            aead_decrypt(
-                self.k.as_ref().unwrap(),
-                self.n,
-                &[],
-                enc_static,
-                &mut static_buf,
-            )
-            .map_err(|_| NoiseError::DecryptionFailed)?;
-            self.n += 1;
-            self.h = mix_hash(&self.h, enc_static);
-
-            let ss_dh =
-                x_only_ecdh(&self.s_priv, &static_buf).map_err(|_| NoiseError::DecryptionFailed)?;
-            let (ck, k) = mix_key(&self.ck, &ss_dh);
-            self.ck = ck;
-            self.k = Some(k);
-            self.n = 0;
-            self.rs_pub = Some(static_buf);
-
-            let enc_epoch = &payload[49..];
-            let mut epoch_buf = [0u8; EPOCH_SIZE];
-            aead_decrypt(
-                self.k.as_ref().unwrap(),
-                self.n,
-                &[],
-                enc_epoch,
-                &mut epoch_buf,
-            )
-            .map_err(|_| NoiseError::DecryptionFailed)?;
-            self.n += 1;
-            self.h = mix_hash(&self.h, enc_epoch);
-
-            Ok((static_buf, epoch_buf))
-        }
-
-        // FIPS: bd08505 noise/handshake.rs:write_message_2()
-        //
-        // Write message 2 for the initiator.
-        //
-        // Returns the number of bytes written to `out`.
-        // Returns `Err` if output buffer is too small.
-        pub fn write_message2(
-            &mut self,
-            responder_ephemeral_secret: &[u8; 32],
-            epoch: &[u8; EPOCH_SIZE],
-            out: &mut [u8],
-        ) -> Result<usize, NoiseError> {
-            let needed = PUBKEY_SIZE + EPOCH_SIZE + TAG_SIZE;
-            if out.len() < needed {
-                return Err(NoiseError::MessageTooLarge {
-                    size: out.len(),
-                    max: needed,
-                });
-            }
-
-            let e_pub = ecdh_pubkey(responder_ephemeral_secret)
-                .map_err(|_| NoiseError::InvalidPublicKey)?;
-            let mut pos = 0;
-
-            out[pos..pos + PUBKEY_SIZE].copy_from_slice(&e_pub);
-            pos += PUBKEY_SIZE;
-            self.h = mix_hash(&self.h, &e_pub);
-
-            let ee_dh = x_only_ecdh(responder_ephemeral_secret, &self.ei_pub)
-                .map_err(|_| NoiseError::DecryptionFailed)?;
-            let (new_ck, k) = mix_key(&self.ck, &ee_dh);
-            self.ck = new_ck;
-            self.k = Some(k);
-            self.n = 0;
-
-            let se_dh = x_only_ecdh(&self.s_priv, &self.ei_pub)
-                .map_err(|_| NoiseError::DecryptionFailed)?;
-            let (new_ck, k) = mix_key(&self.ck, &se_dh);
-            self.ck = new_ck;
-            self.k = Some(k);
-            self.n = 0;
-
-            let enc_len = aead_encrypt(
-                self.k.as_ref().unwrap(),
-                self.n,
-                &[],
-                epoch,
-                &mut out[pos..],
-            )
-            .map_err(|_| NoiseError::EncryptionFailed)?;
-            self.n += 1;
-            self.h = mix_hash(&self.h, &out[pos..pos + enc_len]);
-            pos += enc_len;
-
-            Ok(pos)
-        }
-
-        // FIPS: bd08505 noise/handshake.rs:SymmetricState::split()
-        //
-        // Finalize the handshake, returning transport keys.
-        //
-        // Returns `(k1, k2)` where k1 and k2 are the first and second
-        // 32-byte keys from HKDF-SHA256(ck, ""). The responder must use
-        // **k2 as k_send** and **k1 as k_recv** (swapped from initiator
-        // convention) because `se` uses the same DH inputs as `es`.
-        pub fn finalize(&self) -> ([u8; 32], [u8; 32]) {
-            let hk = Hkdf::<Sha256>::new(Some(&self.ck), &[]);
-            let mut okm = [0u8; 64];
-            hk.expand(&[], &mut okm).expect("hkdf expand failed");
-            let mut k1 = [0u8; 32];
-            let mut k2 = [0u8; 32];
-            k1.copy_from_slice(&okm[..32]);
-            k2.copy_from_slice(&okm[32..]);
-            (k1, k2)
-        }
-    }
-}
-
-pub use responder_pub::NoiseIkResponder;
-
-#[cfg(test)]
-mod responder_tests {
-    use super::*;
-
-    #[test]
-    fn noise_ik_full_handshake_simulation() {
-        // Use fresh random keys to prove correctness with any valid keypair.
-        let (initiator_eph_secret, _) = test_keypair();
-        let (initiator_static_secret, _) = test_keypair();
-        let (responder_static_secret, _) = test_keypair();
-        let (responder_eph_secret, _) = test_keypair();
-        let responder_static_pub = ecdh_pubkey(&responder_static_secret).unwrap();
-        let initiator_static_pub = ecdh_pubkey(&initiator_static_secret).unwrap();
-        let epoch_a = [0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
-        let epoch_b = [0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
-
-        let (mut initiator, _) = NoiseIkInitiator::new(
-            &initiator_eph_secret,
-            &initiator_static_secret,
-            &responder_static_pub,
-        )
-        .unwrap();
-
-        let mut msg1_buf = [0u8; 256];
-        let msg1_len = initiator
-            .write_message1(&initiator_static_pub, &epoch_a, &mut msg1_buf)
-            .unwrap();
-        assert_eq!(msg1_len, 106);
-
-        let mut responder =
-            NoiseIkResponder::new(&responder_static_secret, msg1_buf[..33].try_into().unwrap())
-                .unwrap();
-        let (received_static_pub, received_epoch_a) =
-            responder.read_message1(&msg1_buf[33..msg1_len]).unwrap();
-        assert_eq!(received_static_pub, initiator_static_pub);
-        assert_eq!(received_epoch_a, epoch_a);
-
-        let mut msg2_buf = [0u8; 128];
-        let msg2_len = responder
-            .write_message2(&responder_eph_secret, &epoch_b, &mut msg2_buf)
-            .unwrap();
-        assert_eq!(msg2_len, 57);
-
-        let received_epoch_b = initiator.read_message2(&msg2_buf[..msg2_len]).unwrap();
-        assert_eq!(received_epoch_b, epoch_b);
-
-        let (k_send_i, k_recv_i) = initiator.finalize();
-        let (k1_r, k2_r) = responder.finalize();
-        // NOTE: k_send_i != k2_r — known issue tracked in GitHub.
-        // The IK responder is not used in practice (MCU is always IK initiator).
-        assert_ne!(k_send_i, [0u8; 32]);
-        assert_ne!(k_recv_i, [0u8; 32]);
-        assert_ne!(k1_r, [0u8; 32]);
-        assert_ne!(k2_r, [0u8; 32]);
-    }
-
-    #[test]
-    fn noise_ik_msg2_size() {
-        // Use fresh random keys to prove message size is correct regardless of key material.
-        let (initiator_eph_secret, _) = test_keypair();
-        let (initiator_static_secret, _) = test_keypair();
-        let (responder_static_secret, _) = test_keypair();
-        let (responder_eph_secret, _) = test_keypair();
-        let responder_static_pub = ecdh_pubkey(&responder_static_secret).unwrap();
-        let initiator_static_pub = ecdh_pubkey(&initiator_static_secret).unwrap();
-        let epoch_a = [0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
-        let epoch_b = [0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
-
-        let (mut initiator, _) = NoiseIkInitiator::new(
-            &initiator_eph_secret,
-            &initiator_static_secret,
-            &responder_static_pub,
-        )
-        .unwrap();
-        let mut msg1_buf = [0u8; 256];
-        let msg1_len = initiator
-            .write_message1(&initiator_static_pub, &epoch_a, &mut msg1_buf)
-            .unwrap();
-
-        let mut responder =
-            NoiseIkResponder::new(&responder_static_secret, msg1_buf[..33].try_into().unwrap())
-                .unwrap();
-        responder.read_message1(&msg1_buf[33..msg1_len]).unwrap();
-
-        let mut msg2_buf = [0u8; 128];
-        let msg2_len = responder
-            .write_message2(&responder_eph_secret, &epoch_b, &mut msg2_buf)
-            .unwrap();
-        // 33 (re_pub) + 24 (enc_epoch = 8 + 16 tag) = 57
-        assert_eq!(msg2_len, 57);
-    }
-
-    #[test]
-    fn noise_ik_transport_keys_are_deterministic() {
-        // KNOWN-ANSWER-VECTOR TEST: Uses fixed keys intentionally to verify
-        // that the same inputs always produce the same transport keys. This
-        // is the one test that must NOT use random keys — it tests determinism.
-        let initiator_eph_secret: [u8; 32] = [0x01; 32];
-        let initiator_static_secret: [u8; 32] = [0x11; 32];
-        let responder_static_secret: [u8; 32] = [0x22; 32];
-        let responder_eph_secret: [u8; 32] = [0xAA; 32];
-        let responder_static_pub = ecdh_pubkey(&responder_static_secret).unwrap();
-        let initiator_static_pub = ecdh_pubkey(&initiator_static_secret).unwrap();
-        let epoch_a = [0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
-        let epoch_b = [0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
-
-        let (mut init1, _) = NoiseIkInitiator::new(
-            &initiator_eph_secret,
-            &initiator_static_secret,
-            &responder_static_pub,
-        )
-        .unwrap();
-        let mut msg1_buf = [0u8; 256];
-        let msg1_len = init1
-            .write_message1(&initiator_static_pub, &epoch_a, &mut msg1_buf)
-            .unwrap();
-
-        let mut resp =
-            NoiseIkResponder::new(&responder_static_secret, msg1_buf[..33].try_into().unwrap())
-                .unwrap();
-        resp.read_message1(&msg1_buf[33..msg1_len]).unwrap();
-
-        let mut msg2_buf = [0u8; 128];
-        let msg2_len = resp
-            .write_message2(&responder_eph_secret, &epoch_b, &mut msg2_buf)
-            .unwrap();
-        init1.read_message2(&msg2_buf[..msg2_len]).unwrap();
-
-        let (k_send1, k_recv1) = init1.finalize();
-
-        // Run again with same keys — must produce identical transport keys
-        let (mut init2, _) = NoiseIkInitiator::new(
-            &initiator_eph_secret,
-            &initiator_static_secret,
-            &responder_static_pub,
-        )
-        .unwrap();
-        let mut msg1_buf2 = [0u8; 256];
-        let msg1_len2 = init2
-            .write_message1(&initiator_static_pub, &epoch_a, &mut msg1_buf2)
-            .unwrap();
-
-        let mut resp2 = NoiseIkResponder::new(
-            &responder_static_secret,
-            msg1_buf2[..33].try_into().unwrap(),
-        )
-        .unwrap();
-        resp2.read_message1(&msg1_buf2[33..msg1_len2]).unwrap();
-
-        let mut msg2_buf2 = [0u8; 128];
-        let msg2_len2 = resp2
-            .write_message2(&responder_eph_secret, &epoch_b, &mut msg2_buf2)
-            .unwrap();
-        init2.read_message2(&msg2_buf2[..msg2_len2]).unwrap();
-
-        let (k_send2, k_recv2) = init2.finalize();
-
-        assert_eq!(k_send1, k_send2, "k_send must be deterministic");
-        assert_eq!(k_recv1, k_recv2, "k_recv must be deterministic");
-    }
-
-    #[test]
-    fn ik_se_uses_es_dh_inputs_intentionally() {
-        // The Noise spec IK pattern says:
-        //   initiator se = DH(s_initiator, re_responder)
-        //   responder se = DH(e_responder, rs_initiator)
-        // These are the same DH result because DH(A,B) == DH(B,A).
-        //
-        // Our initiator's read_message2 computes:
-        //   se = DH(e_initiator, rs_responder)
-        // The responder's write_message2 computes:
-        //   se = DH(s_responder, e_initiator) = DH(e_initiator, s_responder)
-        // This is the standard Noise spec se token, NOT a deviation.
-        //
-        // Note: our initiator ALSO computes es = DH(e_init, rs_resp) in
-        // write_message1. So se and es use the same inputs. This is correct
-        // per the Noise spec because:
-        //   initiator es = DH(e_init, rs_resp)
-        //   initiator se = DH(s_init, re_resp) -- but re is not yet known at
-        //                  this point... wait.
-        //
-        // Actually let me re-check. In Noise IK:
-        //   -> e, es, s, ss
-        //   <- e, ee, se
-        //
-        // For the initiator:
-        //   es = DH(e, rs)  -- known: e (just generated), rs (pre-message)
-        //   se = DH(s, re)  -- known: s (our static), re (from msg2)
-        //
-        // For the responder:
-        //   es = DH(s, ei)  -- known: s (our static), ei (from msg1)
-        //   se = DH(e, ri)  -- known: e (just generated), ri (from msg1, encrypted)
-        //
-        // DH(e_init, rs_resp) = DH(rs_resp, e_init) = DH(s_resp, e_init) = responder's es
-        // DH(s_init, re_resp) = DH(re_resp, s_init) = DH(e_resp, ri) = responder's se
-        //
-        // So initiator's es = responder's es (correct, both DH(e, rs))
-        // And initiator's se = responder's se (correct, both DH(s, re))
-        //
-        // Our code in read_message2 does DH(e_init, rs_resp) for se.
-        // But the Noise spec says se = DH(s_init, re_resp).
-        // DH(e_init, rs_resp) ≠ DH(s_init, re_resp) in general.
-        //
-        // HOWEVER, FIPS code does the same thing on both sides, so it
-        // interoperates. Let's prove our test responder (which mirrors FIPS)
-        // produces the same keys as the initiator, with fresh random keys:
-        let (i_eph, _) = test_keypair();
-        let (i_stat, _) = test_keypair();
-        let (r_stat, _) = test_keypair();
-        let (r_eph, _) = test_keypair();
-        let r_pub = ecdh_pubkey(&r_stat).unwrap();
-        let i_pub = ecdh_pubkey(&i_stat).unwrap();
-        let epoch_i = [0x01, 0, 0, 0, 0, 0, 0, 0];
-        let epoch_r = [0x02, 0, 0, 0, 0, 0, 0, 0];
-
-        let (mut init, _) = NoiseIkInitiator::new(&i_eph, &i_stat, &r_pub).unwrap();
-        let mut msg1 = [0u8; 256];
-        let msg1_len = init.write_message1(&i_pub, &epoch_i, &mut msg1).unwrap();
-
-        let mut resp = NoiseIkResponder::new(&r_stat, msg1[..33].try_into().unwrap()).unwrap();
-        let (recv_pub, recv_epoch) = resp.read_message1(&msg1[33..msg1_len]).unwrap();
-        assert_eq!(recv_pub, i_pub);
-        assert_eq!(recv_epoch, epoch_i);
-
-        let mut msg2 = [0u8; 128];
-        let msg2_len = resp.write_message2(&r_eph, &epoch_r, &mut msg2).unwrap();
-
-        // This MUST succeed — if se DH was wrong, decryption would fail
-        let recv_epoch = init.read_message2(&msg2[..msg2_len]).unwrap();
-        assert_eq!(recv_epoch, epoch_r);
-
-        // Transport keys derived from both sides must match
-        let (k_send_i, k_recv_i) = init.finalize();
-        assert_ne!(k_send_i, [0u8; 32]);
-        assert_ne!(k_recv_i, [0u8; 32]);
-        // k_send_i is what initiator uses to encrypt -> responder decrypts
-        // k_recv_i is what responder uses to encrypt -> initiator decrypts
-    }
-
-    #[test]
-    fn se_and_es_produce_different_keys() {
-        // es = DH(e_init, rs_resp) — used in write_message1
-        // se = DH(e_init, rs_resp) in our code — used in read_message2
-        // These are the SAME DH inputs in our implementation!
-        // This means es and se produce the same shared secret.
-        // In the standard Noise spec they would be different:
-        //   es = DH(e, rs)
-        //   se = DH(s, re)
-        //
-        // D3: x-only ECDH (FIPS design choice for Nostr npub compat, see module docs). Both sides do it,
-        // so keys still match. Our test responder mirrors FIPS exactly.
-        let (i_eph, _) = test_keypair();
-        let (r_stat, _) = test_keypair();
-        let r_pub = ecdh_pubkey(&r_stat).unwrap();
-
-        let es_dh = x_only_ecdh(&i_eph, &r_pub).unwrap();
-        // In our read_message2, se also uses DH(e_init, rs_resp):
-        let se_dh = x_only_ecdh(&i_eph, &r_pub).unwrap();
-        assert_eq!(es_dh, se_dh, "es and se use same inputs in our impl");
-
-        // In standard Noise spec, se would use DH(s_init, re_resp):
-        // We can't test this without knowing re_resp, but the point is
-        // that our implementation matches FIPS (which also uses DH(e, rs) for se).
-    }
-
-    #[test]
-    fn noise_ik_with_real_mcu_keys() {
-        // Use the actual MCU secret key to verify pubkey derivation matches
-        // what we see on the MCU (logged via RTT: pub: [02, 63, 56, 96, ...])
-        let mcu_secret: [u8; 32] = [
-            0xac, 0x68, 0xaf, 0x89, 0x46, 0x2e, 0x7e, 0xd2, 0x6f, 0xf6, 0x70, 0xc1, 0x86, 0xb4,
-            0xee, 0xb5, 0x3c, 0x4e, 0x82, 0xd7, 0x2c, 0x8e, 0xf6, 0xce, 0xc4, 0xe6, 0x76, 0xc7,
-            0x84, 0x3f, 0x83, 0x2e,
-        ];
-        let mcu_pub = ecdh_pubkey(&mcu_secret).unwrap();
-        // RTT logged: pub: [02, 63, 56, 96, dc, 5f, 7c, cb, 68, df, 79, 36, 2c, 9e, df, 35,
-        //                 e3, 5e, 61, 6d, 7a, e8, 6f, ce, e2, 68, a2, f7, 49, 45, 2b, 68, 42]
-        assert_eq!(mcu_pub[0], 0x02);
-        assert_eq!(mcu_pub[1], 0x63); // matches RTT log: pub: [02, 63, 56, 96, ...]
-                                      // The exact pubkey depends on k256's compressed encoding — just verify it's valid
-        assert_eq!(mcu_pub.len(), 33);
-    }
-
-    #[test]
-    fn vps_pubkey_is_valid_secp256k1() {
-        let vps_pub: [u8; 33] = [
-            0x02, 0x0e, 0x7a, 0x0d, 0xa0, 0x1a, 0x25, 0x5c, 0xde, 0x10, 0x6a, 0x20, 0x2e, 0xf4,
-            0xf5, 0x73, 0x67, 0x6e, 0xf9, 0xe2, 0x4f, 0x1c, 0x81, 0x76, 0xd0, 0x3a, 0xe8, 0x3a,
-            0x2a, 0x3a, 0x03, 0x7d, 0x21,
-        ];
-        let _pk = PublicKey::from_sec1_bytes(&vps_pub).unwrap();
-    }
-
-    #[test]
-    fn noise_xk_msg2_size() {
-        let (initiator_eph_secret, _) = test_keypair();
-        let (responder_static_secret, _) = test_keypair();
-        let (responder_eph_secret, _) = test_keypair();
-        let responder_static_pub = ecdh_pubkey(&responder_static_secret).unwrap();
-        let epoch = [0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
-
-        let (mut initiator, _) =
-            NoiseXkInitiator::new(&initiator_eph_secret, &[0x11; 32], &responder_static_pub)
-                .unwrap();
-
-        let mut msg1_buf = [0u8; 64];
-        let _msg1_len = initiator.write_message1(&mut msg1_buf).unwrap();
-
-        let mut responder =
-            NoiseXkResponder::new(&responder_static_secret, msg1_buf[..33].try_into().unwrap())
-                .unwrap();
-
-        let mut msg2_buf = [0u8; 128];
-        let msg2_len = responder
-            .write_message2(&responder_eph_secret, &epoch, &mut msg2_buf)
-            .unwrap();
-        // 33 (re_pub) + 24 (enc_epoch) = 57
-        assert_eq!(msg2_len, 57);
-    }
-
-    #[test]
-    fn noise_xk_msg3_size() {
-        let (initiator_eph_secret, _) = test_keypair();
-        let (initiator_static_secret, _) = test_keypair();
-        let (responder_static_secret, _) = test_keypair();
-        let (responder_eph_secret, _) = test_keypair();
-        let responder_static_pub = ecdh_pubkey(&responder_static_secret).unwrap();
-        let initiator_static_pub = ecdh_pubkey(&initiator_static_secret).unwrap();
-        let epoch_a = [0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
-        let epoch_b = [0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
-
-        let (mut initiator, _) = NoiseXkInitiator::new(
-            &initiator_eph_secret,
-            &initiator_static_secret,
-            &responder_static_pub,
-        )
-        .unwrap();
-
-        let mut msg1_buf = [0u8; 64];
-        initiator.write_message1(&mut msg1_buf).unwrap();
-
-        let mut responder =
-            NoiseXkResponder::new(&responder_static_secret, msg1_buf[..33].try_into().unwrap())
-                .unwrap();
-
-        let mut msg2_buf = [0u8; 128];
-        let msg2_len = responder
-            .write_message2(&responder_eph_secret, &epoch_a, &mut msg2_buf)
-            .unwrap();
-        initiator.read_message2(&msg2_buf[..msg2_len]).unwrap();
-
-        let mut msg3_buf = [0u8; 128];
-        let msg3_len = initiator
-            .write_message3(&initiator_static_pub, &epoch_b, &mut msg3_buf)
-            .unwrap();
-        // 49 (enc_static = 33 + 16) + 24 (enc_epoch = 8 + 16) = 73
-        assert_eq!(msg3_len, 73);
-    }
-
-    #[test]
-    fn noise_xk_full_handshake_simulation() {
-        // Use fresh random keys to prove XK works with any valid keypair.
-        let (initiator_eph_secret, _) = test_keypair();
-        let (initiator_static_secret, _) = test_keypair();
-        let (responder_static_secret, _) = test_keypair();
-        let (responder_eph_secret, _) = test_keypair();
-        let responder_static_pub = ecdh_pubkey(&responder_static_secret).unwrap();
-        let initiator_static_pub = ecdh_pubkey(&initiator_static_secret).unwrap();
-        let epoch_a = [0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
-        let epoch_b = [0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
-
-        let (mut initiator, _) = NoiseXkInitiator::new(
-            &initiator_eph_secret,
-            &initiator_static_secret,
-            &responder_static_pub,
-        )
-        .unwrap();
-
-        let mut msg1_buf = [0u8; 64];
-        let msg1_len = initiator.write_message1(&mut msg1_buf).unwrap();
-        assert_eq!(msg1_len, 33);
-
-        let mut responder =
-            NoiseXkResponder::new(&responder_static_secret, msg1_buf[..33].try_into().unwrap())
-                .unwrap();
-
-        let mut msg2_buf = [0u8; 128];
-        let msg2_len = responder
-            .write_message2(&responder_eph_secret, &epoch_a, &mut msg2_buf)
-            .unwrap();
-        assert_eq!(msg2_len, 57);
-
-        let received_epoch_a = initiator.read_message2(&msg2_buf[..msg2_len]).unwrap();
-        assert_eq!(received_epoch_a, epoch_a);
-
-        let mut msg3_buf = [0u8; 128];
-        let msg3_len = initiator
-            .write_message3(&initiator_static_pub, &epoch_b, &mut msg3_buf)
-            .unwrap();
-        // 49 (enc_static = 33 + 16) + 24 (enc_epoch = 8 + 16) = 73
-        assert_eq!(msg3_len, 73);
-
-        let (received_static_pub, received_epoch_b) =
-            responder.read_message3(&msg3_buf[..msg3_len]).unwrap();
-        assert_eq!(received_static_pub, initiator_static_pub);
-        assert_eq!(received_epoch_b, epoch_b);
-
-        let (k_send_i, k_recv_i) = initiator.finalize();
-        let (k_recv_r, k_send_r) = responder.finalize();
-
-        assert_eq!(k_send_i, k_recv_r, "initiator send == responder recv");
-        assert_eq!(k_recv_i, k_send_r, "initiator recv == responder send");
-    }
-
-    #[test]
-    fn noise_xk_transport_keys_are_deterministic() {
-        // KNOWN-ANSWER-VECTOR TEST: Uses fixed keys intentionally to verify
-        // that the same inputs always produce the same transport keys.
-        let initiator_eph_secret: [u8; 32] = [0x01; 32];
-        let initiator_static_secret: [u8; 32] = [0x11; 32];
-        let responder_static_secret: [u8; 32] = [0x22; 32];
-        let responder_eph_secret: [u8; 32] = [0xAA; 32];
-        let responder_static_pub = ecdh_pubkey(&responder_static_secret).unwrap();
-        let initiator_static_pub = ecdh_pubkey(&initiator_static_secret).unwrap();
-        let epoch_a = [0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
-        let epoch_b = [0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
-
-        let (mut init1, _) = NoiseXkInitiator::new(
-            &initiator_eph_secret,
-            &initiator_static_secret,
-            &responder_static_pub,
-        )
-        .unwrap();
-        let mut msg1_buf = [0u8; 64];
-        init1.write_message1(&mut msg1_buf).unwrap();
-
-        let mut resp =
-            NoiseXkResponder::new(&responder_static_secret, msg1_buf[..33].try_into().unwrap())
-                .unwrap();
-        let mut msg2_buf = [0u8; 128];
-        let msg2_len = resp
-            .write_message2(&responder_eph_secret, &epoch_a, &mut msg2_buf)
-            .unwrap();
-        init1.read_message2(&msg2_buf[..msg2_len]).unwrap();
-        let mut msg3_buf = [0u8; 128];
-        init1
-            .write_message3(&initiator_static_pub, &epoch_b, &mut msg3_buf)
-            .unwrap();
-
-        let (k_send1, k_recv1) = init1.finalize();
-
-        let (mut init2, _) = NoiseXkInitiator::new(
-            &initiator_eph_secret,
-            &initiator_static_secret,
-            &responder_static_pub,
-        )
-        .unwrap();
-        let mut msg1_buf2 = [0u8; 64];
-        init2.write_message1(&mut msg1_buf2).unwrap();
-
-        let mut resp2 = NoiseXkResponder::new(
-            &responder_static_secret,
-            msg1_buf2[..33].try_into().unwrap(),
-        )
-        .unwrap();
-        let mut msg2_buf2 = [0u8; 128];
-        let msg2_len2 = resp2
-            .write_message2(&responder_eph_secret, &epoch_a, &mut msg2_buf2)
-            .unwrap();
-        init2.read_message2(&msg2_buf2[..msg2_len2]).unwrap();
-        let mut msg3_buf2 = [0u8; 128];
-        init2
-            .write_message3(&initiator_static_pub, &epoch_b, &mut msg3_buf2)
-            .unwrap();
-
-        let (k_send2, k_recv2) = init2.finalize();
-
-        assert_eq!(k_send1, k_send2, "k_send must be deterministic");
-        assert_eq!(k_recv1, k_recv2, "k_recv must be deterministic");
-    }
-
-    #[test]
-    fn noise_xk_ck_matches_step_by_step() {
-        // Use fresh random keys to verify step-by-step DH agreement.
-        let (initiator_eph_secret, _) = test_keypair();
-        let (initiator_static_secret, _) = test_keypair();
-        let (responder_static_secret, _) = test_keypair();
-        let (responder_eph_secret, _) = test_keypair();
-        let responder_static_pub = ecdh_pubkey(&responder_static_secret).unwrap();
-        let initiator_static_pub = ecdh_pubkey(&initiator_static_secret).unwrap();
-        let initiator_eph_pub = ecdh_pubkey(&initiator_eph_secret).unwrap();
-        let responder_eph_pub = ecdh_pubkey(&responder_eph_secret).unwrap();
-
-        let h0 = hash_one(PROTOCOL_NAME_XK);
-        let mut ck_i = h0;
-        let mut ck_r = h0;
-        let mut h_i = h0;
-        let mut h_r = h0;
-
-        let norm_rs = parity_normalize(&responder_static_pub);
-        h_i = mix_hash(&h_i, &norm_rs);
-        let norm_s = parity_normalize(&responder_static_pub);
-        h_r = mix_hash(&h_r, &norm_s);
-        assert_eq!(h_i, h_r, "h after pre-message");
-
-        h_i = mix_hash(&h_i, &initiator_eph_pub);
-        h_r = mix_hash(&h_r, &initiator_eph_pub);
-        assert_eq!(h_i, h_r, "h after e");
-
-        let es_i = x_only_ecdh(&initiator_eph_secret, &responder_static_pub).unwrap();
-        let es_r = x_only_ecdh(&responder_static_secret, &initiator_eph_pub).unwrap();
-        assert_eq!(es_i, es_r, "es DH");
-        let (ck_i_1, k_i_1) = mix_key(&ck_i, &es_i);
-        let (ck_r_1, k_r_1) = mix_key(&ck_r, &es_r);
-        assert_eq!(ck_i_1, ck_r_1, "ck after es");
-        assert_eq!(k_i_1, k_r_1, "k after es");
-        ck_i = ck_i_1;
-        ck_r = ck_r_1;
-
-        h_i = mix_hash(&h_i, &responder_eph_pub);
-        h_r = mix_hash(&h_r, &responder_eph_pub);
-        assert_eq!(h_i, h_r, "h after re");
-
-        let ee_i = x_only_ecdh(&initiator_eph_secret, &responder_eph_pub).unwrap();
-        let ee_r = x_only_ecdh(&responder_eph_secret, &initiator_eph_pub).unwrap();
-        assert_eq!(ee_i, ee_r, "ee DH");
-        let (ck_i_2, k_i_2) = mix_key(&ck_i, &ee_i);
-        let (ck_r_2, k_r_2) = mix_key(&ck_r, &ee_r);
-        assert_eq!(ck_i_2, ck_r_2, "ck after ee");
-        assert_eq!(k_i_2, k_r_2, "k after ee");
-        ck_i = ck_i_2;
-        ck_r = ck_r_2;
-
-        let se_i = x_only_ecdh(&initiator_static_secret, &responder_eph_pub).unwrap();
-        let se_r = x_only_ecdh(&responder_eph_secret, &initiator_static_pub).unwrap();
-        assert_eq!(se_i, se_r, "se DH");
-        let (ck_i_3, k_i_3) = mix_key(&ck_i, &se_i);
-        let (ck_r_3, k_r_3) = mix_key(&ck_r, &se_r);
-        assert_eq!(ck_i_3, ck_r_3, "ck after se");
-        assert_eq!(k_i_3, k_r_3, "k after se");
-        ck_i = ck_i_3;
-        ck_r = ck_r_3;
-
-        let (k_send_i, k_recv_i) = {
-            let hk = Hkdf::<Sha256>::new(Some(&ck_i), &[]);
-            let mut okm = [0u8; 64];
-            hk.expand(&[], &mut okm).unwrap();
-            let mut k1 = [0u8; 32];
-            let mut k2 = [0u8; 32];
-            k1.copy_from_slice(&okm[..32]);
-            k2.copy_from_slice(&okm[32..]);
-            (k1, k2)
-        };
-        let (k_send_r, k_recv_r) = {
-            let hk = Hkdf::<Sha256>::new(Some(&ck_r), &[]);
-            let mut okm = [0u8; 64];
-            hk.expand(&[], &mut okm).unwrap();
-            let mut k1 = [0u8; 32];
-            let mut k2 = [0u8; 32];
-            k1.copy_from_slice(&okm[..32]);
-            k2.copy_from_slice(&okm[32..]);
-            (k1, k2)
-        };
-        assert_eq!(k_send_i, k_send_r, "final k_send");
-        assert_eq!(k_recv_i, k_recv_r, "final k_recv");
-    }
-
-    #[test]
-    fn test_xk_initiator_responder_full_handshake() {
-        const SECRET_A: [u8; 32] = [0xAA; 32];
-        const SECRET_B: [u8; 32] = [0xBB; 32];
-
-        let responder_static_pub = ecdh_pubkey(&SECRET_B).unwrap();
-        let initiator_static_pub = ecdh_pubkey(&SECRET_A).unwrap();
-        let epoch2 = [0x11; EPOCH_SIZE];
-        let epoch3 = [0x22; EPOCH_SIZE];
-
-        let (mut initiator, initiator_eph_pub) =
-            NoiseXkInitiator::new(&SECRET_A, &SECRET_A, &responder_static_pub).unwrap();
-        let mut msg1 = [0u8; crate::fsp::XK_HANDSHAKE_MSG1_SIZE];
-        let msg1_len = initiator.write_message1(&mut msg1).unwrap();
-        assert_eq!(msg1_len, crate::fsp::XK_HANDSHAKE_MSG1_SIZE);
-
-        let mut responder = NoiseXkResponder::new(&SECRET_B, &initiator_eph_pub).unwrap();
-        let mut msg2 = [0u8; crate::fsp::XK_HANDSHAKE_MSG2_SIZE];
-        let msg2_len = responder
-            .write_message2(&SECRET_B, &epoch2, &mut msg2)
-            .unwrap();
-        assert_eq!(msg2_len, crate::fsp::XK_HANDSHAKE_MSG2_SIZE);
-        assert_eq!(initiator.read_message2(&msg2).unwrap(), epoch2);
-
-        let mut msg3 = [0u8; crate::fsp::XK_HANDSHAKE_MSG3_SIZE];
-        let msg3_len = initiator
-            .write_message3(&initiator_static_pub, &epoch3, &mut msg3)
-            .unwrap();
-        assert_eq!(msg3_len, crate::fsp::XK_HANDSHAKE_MSG3_SIZE);
-
-        let (received_static, received_epoch) = responder.read_message3(&msg3).unwrap();
-        assert_eq!(received_static, initiator_static_pub);
-        assert_eq!(received_epoch, epoch3);
-
-        let (k_send_i, k_recv_i) = initiator.finalize();
-        let (k_recv_r, k_send_r) = responder.finalize();
-        assert_eq!(k_send_i, k_recv_r);
-        assert_eq!(k_recv_i, k_send_r);
-    }
-
-    #[test]
-    fn test_xk_message_sizes() {
-        const SECRET_A: [u8; 32] = [0xAA; 32];
-        const SECRET_B: [u8; 32] = [0xBB; 32];
-
-        let responder_static_pub = ecdh_pubkey(&SECRET_B).unwrap();
-        let initiator_static_pub = ecdh_pubkey(&SECRET_A).unwrap();
-
-        let (mut initiator, initiator_eph_pub) =
-            NoiseXkInitiator::new(&SECRET_A, &SECRET_A, &responder_static_pub).unwrap();
-
-        let mut msg1 = [0u8; crate::fsp::XK_HANDSHAKE_MSG1_SIZE];
-        let msg1_len = initiator.write_message1(&mut msg1).unwrap();
-        assert_eq!(msg1_len, crate::fsp::XK_HANDSHAKE_MSG1_SIZE);
-        assert_eq!(msg1_len, 33);
-
-        let mut responder = NoiseXkResponder::new(&SECRET_B, &initiator_eph_pub).unwrap();
-        let mut msg2 = [0u8; crate::fsp::XK_HANDSHAKE_MSG2_SIZE];
-        let msg2_len = responder
-            .write_message2(&SECRET_B, &[0x33; EPOCH_SIZE], &mut msg2)
-            .unwrap();
-        assert_eq!(msg2_len, crate::fsp::XK_HANDSHAKE_MSG2_SIZE);
-        assert_eq!(msg2_len, 57);
-
-        initiator.read_message2(&msg2).unwrap();
-        let mut msg3 = [0u8; crate::fsp::XK_HANDSHAKE_MSG3_SIZE];
-        let msg3_len = initiator
-            .write_message3(&initiator_static_pub, &[0x44; EPOCH_SIZE], &mut msg3)
-            .unwrap();
-        assert_eq!(msg3_len, crate::fsp::XK_HANDSHAKE_MSG3_SIZE);
-        assert_eq!(msg3_len, 73);
-    }
-
-    #[test]
-    fn test_xk_wrong_remote_key() {
-        const SECRET_A: [u8; 32] = [0xAA; 32];
-        const SECRET_B: [u8; 32] = [0xBB; 32];
-        const SECRET_C: [u8; 32] = [0xCC; 32];
-
-        let wrong_responder_pub = ecdh_pubkey(&SECRET_C).unwrap();
-        let (mut initiator, initiator_eph_pub) =
-            NoiseXkInitiator::new(&SECRET_A, &SECRET_A, &wrong_responder_pub).unwrap();
-
-        let mut msg1 = [0u8; crate::fsp::XK_HANDSHAKE_MSG1_SIZE];
-        initiator.write_message1(&mut msg1).unwrap();
-
-        let mut responder = NoiseXkResponder::new(&SECRET_B, &initiator_eph_pub).unwrap();
-        let mut msg2 = [0u8; crate::fsp::XK_HANDSHAKE_MSG2_SIZE];
-        responder
-            .write_message2(&SECRET_B, &[0x55; EPOCH_SIZE], &mut msg2)
-            .unwrap();
-
-        assert!(initiator.read_message2(&msg2).is_err());
-    }
-
-    #[test]
-    fn test_xk_replay_msg1() {
-        const SECRET_A: [u8; 32] = [0xAA; 32];
-        const SECRET_B: [u8; 32] = [0xBB; 32];
-
-        let responder_static_pub = ecdh_pubkey(&SECRET_B).unwrap();
-        let initiator_static_pub = ecdh_pubkey(&SECRET_A).unwrap();
-
-        let (mut initiator, initiator_eph_pub) =
-            NoiseXkInitiator::new(&SECRET_A, &SECRET_A, &responder_static_pub).unwrap();
-        let mut msg1 = [0u8; crate::fsp::XK_HANDSHAKE_MSG1_SIZE];
-        initiator.write_message1(&mut msg1).unwrap();
-
-        let mut responder = NoiseXkResponder::new(&SECRET_B, &initiator_eph_pub).unwrap();
-        let mut msg2 = [0u8; crate::fsp::XK_HANDSHAKE_MSG2_SIZE];
-        responder
-            .write_message2(&SECRET_B, &[0x66; EPOCH_SIZE], &mut msg2)
-            .unwrap();
-        initiator.read_message2(&msg2).unwrap();
-
-        let mut msg3 = [0u8; crate::fsp::XK_HANDSHAKE_MSG3_SIZE];
-        initiator
-            .write_message3(&initiator_static_pub, &[0x77; EPOCH_SIZE], &mut msg3)
-            .unwrap();
-
-        assert!(responder.read_message3(&msg3).is_ok());
-        assert!(responder.read_message3(&msg3).is_err());
-    }
-
-    #[test]
-    fn test_xk_wrong_order() {
-        const SECRET_A: [u8; 32] = [0xAA; 32];
-        const SECRET_B: [u8; 32] = [0xBB; 32];
-
-        let initiator_eph_pub = ecdh_pubkey(&SECRET_A).unwrap();
-        let mut responder = NoiseXkResponder::new(&SECRET_B, &initiator_eph_pub).unwrap();
-        let initiator_static_pub = ecdh_pubkey(&SECRET_A).unwrap();
-        let mut msg3 = [0u8; crate::fsp::XK_HANDSHAKE_MSG3_SIZE];
-
-        let enc_static_len = aead_encrypt(
-            responder.k.as_ref().unwrap(),
-            0,
-            &[],
-            &initiator_static_pub,
-            &mut msg3,
-        )
-        .unwrap();
-        assert_eq!(enc_static_len, PUBKEY_SIZE + TAG_SIZE);
-
-        let enc_epoch_len = aead_encrypt(
-            responder.k.as_ref().unwrap(),
-            1,
-            &[],
-            &[0x99; EPOCH_SIZE],
-            &mut msg3[enc_static_len..],
-        )
-        .unwrap();
-        assert_eq!(enc_epoch_len, EPOCH_SIZE + TAG_SIZE);
-
-        assert_eq!(
-            responder.read_message3(&msg3),
-            Err(NoiseError::InvalidState)
-        );
-    }
-
-    // --- Noise XX Tests ---
-
-    #[test]
-    fn noise_xx_msg1_is_ephemeral_only() {
-        let (eph, _) = test_keypair();
-        let (stat, _) = test_keypair();
-        let (mut init, e_pub) = NoiseXxInitiator::new(&eph, &stat).unwrap();
-
-        let mut msg1 = [0u8; 256];
-        let len = init.write_message1(&mut msg1).unwrap();
-        assert_eq!(len, XX_HANDSHAKE_MSG1_SIZE);
-        assert_eq!(&msg1[..PUBKEY_SIZE], &e_pub);
-    }
-
-    #[test]
-    fn noise_xx_full_round_trip() {
-        let (i_eph, _) = test_keypair();
-        let (i_stat, i_pub) = test_keypair();
-        let (r_stat, r_pub) = test_keypair();
-        let (r_eph, _) = test_keypair();
-        let epoch_i = [0x01, 0, 0, 0, 0, 0, 0, 0];
-        let epoch_r = [0x02, 0, 0, 0, 0, 0, 0, 0];
-
-        let (mut init, _) = NoiseXxInitiator::new(&i_eph, &i_stat).unwrap();
-        let mut resp = NoiseXxResponder::new(&r_stat).unwrap();
-
-        let mut msg1 = [0u8; 256];
-        let msg1_len = init.write_message1(&mut msg1).unwrap();
-        assert_eq!(msg1_len, XX_HANDSHAKE_MSG1_SIZE);
-
-        resp.read_message1(&msg1[..msg1_len]).unwrap();
-
-        let mut msg2 = [0u8; 256];
-        let msg2_len = resp.write_message2(&r_eph, &epoch_r, &mut msg2).unwrap();
-        assert_eq!(msg2_len, XX_HANDSHAKE_MSG2_SIZE);
-
-        let (recv_r_pub, recv_epoch_r) = init.read_message2(&msg2[..msg2_len]).unwrap();
-        assert_eq!(recv_r_pub, r_pub);
-        assert_eq!(recv_epoch_r, epoch_r);
-
-        let mut msg3 = [0u8; 256];
-        let msg3_len = init.write_message3(&i_pub, &epoch_i, &mut msg3).unwrap();
-        assert_eq!(msg3_len, XX_HANDSHAKE_MSG3_SIZE);
-
-        let (recv_i_pub, recv_epoch_i) = resp.read_message3(&msg3[..msg3_len]).unwrap();
-        assert_eq!(recv_i_pub, i_pub);
-        assert_eq!(recv_epoch_i, epoch_i);
-
-        let (c1_i, c2_i) = init.finalize();
-        let (c1_r, c2_r) = resp.finalize();
-
-        assert_eq!(c1_i, c1_r, "both sides derive same c1 (init->resp)");
-        assert_eq!(c2_i, c2_r, "both sides derive same c2 (resp->init)");
-        assert_ne!(c1_i, c2_i, "c1 != c2");
-    }
-
-    #[test]
-    fn noise_xx_keys_are_deterministic() {
-        let i_eph: [u8; 32] = [0x01; 32];
-        let i_stat: [u8; 32] = [0x11; 32];
-        let r_stat: [u8; 32] = [0x22; 32];
-        let r_eph: [u8; 32] = [0xAA; 32];
-        let epoch_i = [0x01, 0, 0, 0, 0, 0, 0, 0];
-        let epoch_r = [0x02, 0, 0, 0, 0, 0, 0, 0];
-
-        let i_pub = ecdh_pubkey(&i_stat).unwrap();
-
-        let (mut init1, _) = NoiseXxInitiator::new(&i_eph, &i_stat).unwrap();
-        let mut resp1 = NoiseXxResponder::new(&r_stat).unwrap();
-
-        let mut m1 = [0u8; 256];
-        let l1 = init1.write_message1(&mut m1).unwrap();
-        resp1.read_message1(&m1[..l1]).unwrap();
-
-        let mut m2 = [0u8; 256];
-        let l2 = resp1.write_message2(&r_eph, &epoch_r, &mut m2).unwrap();
-        init1.read_message2(&m2[..l2]).unwrap();
-
-        let mut m3 = [0u8; 256];
-        let l3 = init1.write_message3(&i_pub, &epoch_i, &mut m3).unwrap();
-        resp1.read_message3(&m3[..l3]).unwrap();
-
-        let (ks1, kr1) = init1.finalize();
-
-        let (mut init2, _) = NoiseXxInitiator::new(&i_eph, &i_stat).unwrap();
-        let mut resp2 = NoiseXxResponder::new(&r_stat).unwrap();
-
-        let mut m1b = [0u8; 256];
-        let l1b = init2.write_message1(&mut m1b).unwrap();
-        resp2.read_message1(&m1b[..l1b]).unwrap();
-
-        let mut m2b = [0u8; 256];
-        let l2b = resp2.write_message2(&r_eph, &epoch_r, &mut m2b).unwrap();
-        init2.read_message2(&m2b[..l2b]).unwrap();
-
-        let mut m3b = [0u8; 256];
-        let l3b = init2.write_message3(&i_pub, &epoch_i, &mut m3b).unwrap();
-        resp2.read_message3(&m3b[..l3b]).unwrap();
-
-        let (ks2, kr2) = init2.finalize();
-
-        assert_eq!(ks1, ks2, "send key deterministic");
-        assert_eq!(kr1, kr2, "recv key deterministic");
-    }
-
-    #[test]
-    fn noise_xx_wrong_msg1_size_rejected() {
-        let (stat, _) = test_keypair();
-        let mut resp = NoiseXxResponder::new(&stat).unwrap();
-        assert_eq!(
-            resp.read_message1(&[0u8; 32]),
-            Err(NoiseError::InvalidMessage)
-        );
-    }
-
-    #[test]
-    fn noise_xx_negotiation_payload_roundtrip() {
-        let (i_eph, _) = test_keypair();
-        let (i_stat, _) = test_keypair();
-        let (r_stat, _) = test_keypair();
-        let (r_eph, _) = test_keypair();
-        let epoch_i = [0x01, 0, 0, 0, 0, 0, 0, 0];
-        let epoch_r = [0x02, 0, 0, 0, 0, 0, 0, 0];
-        let i_pub = ecdh_pubkey(&i_stat).unwrap();
-
-        let (mut init, _) = NoiseXxInitiator::new(&i_eph, &i_stat).unwrap();
-        let mut resp = NoiseXxResponder::new(&r_stat).unwrap();
-
-        let mut m1 = [0u8; 256];
-        let l1 = init.write_message1(&mut m1).unwrap();
-        resp.read_message1(&m1[..l1]).unwrap();
-
-        let mut m2 = [0u8; 256];
-        let l2 = resp.write_message2(&r_eph, &epoch_r, &mut m2).unwrap();
-        init.read_message2(&m2[..l2]).unwrap();
-
-        let mut m3 = [0u8; 256];
-        let l3 = init.write_message3(&i_pub, &epoch_i, &mut m3).unwrap();
-        resp.read_message3(&m3[..l3]).unwrap();
-
-        let payload = [0xDE, 0xAD, 0xBE, 0xEF, 0xCA, 0xFE];
-
-        let mut enc = [0u8; 256];
-        let enc_len = init.encrypt_payload(&payload, &mut enc).unwrap();
-
-        let mut dec = [0u8; 256];
-        let dec_len = resp.decrypt_payload(&enc[..enc_len], &mut dec).unwrap();
-
-        assert_eq!(&dec[..dec_len], &payload);
-    }
-}
+//! All Noise types, constants, and helper functions now live in the
+//! standalone `fips-noise` crate. This module re-exports everything
+//! for backward compatibility with existing `crate::noise::` imports.
+pub use fips_noise::*;
